### 2024-11-12 #7465 @editorialbot
[PRE REVIEW]: NeuroAnalyzer: Julia toolbox for analyzing neurophysiological data
**Submitting author:** @AdamWysokinski (<a href="http://orcid.org/0000-0002-6159-6579">Adam Wysokiński</a>)
**Repository:** https://github.com/JuliaHealth/NeuroAnalyzer.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.24.11
