### 2024-11-12 #7465 @AdamWysokinski
> For now, we will need to waitlist this paper and process it as the queue reduces. Thanks for your patience!

@arfon Thank you for letting me know. No problem, I'm not in a hurry. Just tell me if there is anything more that I should do now, since I'm new to the JOSS reviewing process.
