+++ src/NeuroAnalyzer.jl
@@ -323,6 +323,7 @@ include("io/import_dat.jl")
 include("io/import_digitrack.jl")
 include("io/import_edf.jl")
 include("io/import_fiff.jl")
+include("io/import_ft.jl")
 include("io/import_set.jl")
 include("io/load_locs.jl")
 include("io/export_locs.jl")
