+++ src/io/import_ft.jl
@@ -1,0 +1,169 @@
+export import_ft
+
+"""
+    import_ft(file_name; <keyword arguments>)
+
+Load FieldTrip file (.mat) and return `NeuroAnalyzer.NEURO` object.
+
+# Arguments
+
+- `file_name::String`: name of the file to load
+- `type::Symbol=:eeg`: type of imported data
+- `detect_type::Bool=true`: detect channel type based on its label
+
+# Returns
+
+- `obj::NeuroAnalyzer.NEURO`
+"""
+function import_ft(file_name::String; type::Symbol=:eeg, detect_type::Bool=true)::NeuroAnalyzer.NEURO
+
+    _wip()
+
+    _check_var(type, [:eeg, :meg, :nirs], "type")
+    @assert type === :eeg "Currently only EEG datatype is supported."
+    data_type = String(type)
+
+    @assert isfile(file_name) "File $file_name cannot be loaded."
+
+    file_type = "FT"
+
+    dataset = matread(file_name)
+    @assert length(keys(dataset)) == 1 "Datasets containing > 1 object are not supported; if you have such a file, please send it to adam.wysokinski@neuroanalyzer.org"
+    _info("Reading object: $(String.(keys(dataset))[1])")
+    dataset = dataset[String.(keys(dataset))[1]]
+
+    @assert "hdr" in keys(dataset) "Dataset does not contain header."
+    hdr = dataset["hdr"]
+    @assert "nChans" in keys(hdr) "Dataset header does not contain nChans field."
+    ch_n = Int64(hdr["nChans"])
+    @assert "Fs" in keys(hdr) "Dataset header does not contain Fs field."
+    sampling_rate = round(Int64, hdr["Fs"])
+
+# keys(dataset["cfg"]) = ["detrend", "paddir", "bpfiltdir", "bpfiltord", "usefftfilt", "groupchans", "refmethod", "leadfield", "removemcg", "bsfiltdir", "showlogo", "bpinstabilityfix", "dataformat", "plotfiltresp", "bsfiltdev", "bpfilter", "dftfreq", "method", "reref", "dftfilter", "polyorder", "bsfiltord", "lpfiltwintype", "bsfiltwintype", "bpfiltdf", "keepcfg", "padding", "baselinewindow", "lpfilter", "continuous", "hilbert", "updatesens", "hpfiltdev", "polyremoval", "callinfo", "rectify", "dataset", "headerformat", "lpfiltord", "medianfiltord", "refchannel", "hpinstabilityfix", "checkstring", "bpfiltwintype", "headerfile", "resample", "lpfiltdir", "bsfilter", "hpfiltord", "implicitref", "bsfiltdf", "previous", "bpfilttype", "lpfiltdev", "feedback", "medianfilter", "version", "subspace", "custom", "bsinstabilityfix", "hpfiltdf", "channel", "hpfiltwintype", "absdiff", "bsfilttype", "derivative", "conv", "standardize", "lpfiltdf", "montage", "removeeog", "hpfiltdir", "dftinvert", "precision", "hpfilter", "trl", "tracktimeinfo", "lpfilttype", "checkpath", "toolbox", "denoise", "boxcar", "hpfilttype", "lpinstabilityfix", "bpfiltdev", "outputfilepresent", "demean", "trackmeminfo", "padtype", "keepprevious", "datafile"]
+
[…]
