+++ src/io/import_ft.jl
@@ -21,20 +21,19 @@ function import_ft(file_name::String; type::Symbol=:eeg, detect_type::Bool=true)
 
     _check_var(type, [:eeg, :meg, :nirs], "type")
     @assert type === :eeg "Currently only EEG datatype is supported."
+    file_type = "FT"
     data_type = String(type)
 
     @assert isfile(file_name) "File $file_name cannot be loaded."
-
-    file_type = "FT"
-
     dataset = matread(file_name)
+
     @assert length(keys(dataset)) == 1 "Datasets containing > 1 object are not supported; if you have such a file, please send it to adam.wysokinski@neuroanalyzer.org"
-    _info("Reading object: $(String.(keys(dataset))[1])")
-    dataset = dataset[String.(keys(dataset))[1]]
+    _info("Reading object: $(string.(keys(dataset))[1])")
+    dataset = dataset[string.(keys(dataset))[1]]
 
     @assert "cfg" in keys(dataset) "Dataset does not contain cfg field."
     cfg = dataset["cfg"]
-    @assert "hdr" in keys(dataset) "Dataset does not contain header field."
+    @assert "hdr" in keys(dataset) "Dataset does not contain hdr field."
     hdr = dataset["hdr"]
     @assert "nChans" in keys(hdr) "Dataset header does not contain nChans field."
     ch_n = Int64(hdr["nChans"])
@@ -43,7 +42,7 @@ function import_ft(file_name::String; type::Symbol=:eeg, detect_type::Bool=true)
 
     # get channel labels, types and units
     if length(hdr["label"][:]) > 0 && length(vec(hdr["label"])) == ch_n
-        clabels = String.(hdr["label"][:])
+        clabels = string.(hdr["label"][:])
     else
         clabels = String[]
         for idx in 1:ch_n
@@ -51,41 +50,31 @@ function import_ft(file_name::String; type::Symbol=:eeg, detect_type::Bool=true)
         end
     end
     clabels = _clean_labels(clabels)
+
     if detect_type
         ch_type = _set_channel_types(clabels, "eeg")
         if "chanunit" in keys(hdr)
-            units = String.(hdr["chanunit"][:])
+            units = string.(hdr["chanunit"][:])
             units = replace.(units, "uV"=>"μV")
         else
             units = [_ch_units(ch_type[idx]) for idx in 1:ch_n]
         end
     else
         if "chantype" in keys(hdr)
-            ch_type = String.(hdr["chantype"][:])
+            ch_type = string.(hdr["chantype"][:])
         else
             ch_type = repeat(["eeg"], ch_n)
         end
         if "chanunit" in keys(hdr)
-            units = String.(hdr["chanunit"][:])
+            units = string.(hdr["chanunit"][:])
             units = replace.(units, "uV"=>"μV")
         else
             units = repeat(["μV"], ch_n)
         end
     end
 
-    # TO DO: get referencing
-    if "reref" in keys(dataset["cfg"]) && dataset["cfg"]["reref"] != "no"
-        _info("Embedded referencing is not supported; if you have such a file, please send it to adam.wysokinski@neuroanalyzer.org")
-    else
-        ref = _detect_montage(clabels, ch_type, data_type)
-    end
-
     @assert "trial" in keys(dataset) "Dataset does not contain trial field."
-    if "trial" in keys(dataset)
-        ep_n = size(dataset["trial"], 2)
-    else
-        ep_n = 1
-    end
+    ep_n = "trial" in keys(dataset) ? size(dataset["trial"], 2) : 1
     ep_len = size(dataset["trial"][1], 2)
     data = zeros(ch_n, ep_len, ep_n)
     for idx in 1:ep_n
[…]
