+++ src/utils/time.jl
@@ -1,5 +1,7 @@
 export t2s
 export s2t
+export markers_s2t
+export markers_s2t!
 
 """
     t2s(t, fs)
@@ -82,5 +84,97 @@ Convert time in samples to seconds.
 function s2t(obj::NeuroAnalyzer.NEURO; s::Int64)::Float64
 
     return round(s / sr(obj), digits=2)
+
+end
+
+"""
+    markers_s2t(m; <keyword arguments>)
+
+Convert markers start and length from samples to seconds.
+
+# Arguments
+
+- `m::DataFrame`: markers
+- `fs::Int64`: sampling rate
+
+# Returns
+
+- `m_new::DataFrame`
+"""
+function markers_s2t(m::DataFrame; fs::Int64)::DataFrame
+
+    m_new = deepcopy(m)
+    m_new[:, :start] = round.(m[:, :start] ./ fs, digits=2)
+    m_new[:, :length] = round.(m[:, :length] ./ fs, digits=2)
+
+    return m_new
+
+end
+
+"""
+    markers_s2t!(m; <keyword arguments>)
+
+Convert markers start and length from samples to seconds.
+
+# Arguments
+
+- `m::DataFrame`: markers
+- `fs::Int64`: sampling rate
+
+# Returns
+
+Nothing
+"""
+function markers_s2t!(m::DataFrame; fs::Int64)::Nothing
+
+    m[:, :start] = round.(m[:, :start] ./ fs, digits=2)
+    m[:, :length] = round.(m[:, :length] ./ fs, digits=2)
+
+    return nothing
+
+end
+
+"""
+    markers_s2t(obj; <keyword arguments>)
+
+Convert markers start and length from samples to seconds.
+
+# Arguments
+
+- `obj::NeuroAnalyzer.NEURO`
+
+# Returns
+
+- `m_new::DataFrame`
+"""
+function markers_s2t(obj::NeuroAnalyzer.NEURO)::DataFrame
+
+    m_new = deepcopy(obj.markers)
+    m_new[:, :start] = round.(m_new[:, :start] ./ sr(obj), digits=2)
+    m_new[:, :length] = round.(m_new[:, :length] ./ sr(obj), digits=2)
+
+    return m_new
+
+end
+
+"""
+    markers_s2t!(obj; <keyword arguments>)
+
+Convert markers start and length from samples to seconds.
+
+# Arguments
+
+- `obj::NeuroAnalyzer.NEURO`
+
+# Returns
+
+Nothing
+"""
+function markers_s2t!(obj::NeuroAnalyzer.NEURO)::Nothing
+
+    obj.markers[:, :start] = round.(obj.markers[:, :start] ./ sr(obj), digits=2)
+    obj.markers[:, :length] = round.(obj.markers[:, :length] ./ sr(obj), digits=2)
+
+    return nothing
 
 end
