+++ src/io/import_ft.jl
@@ -9,18 +9,22 @@ Load FieldTrip file (.mat) and return `NeuroAnalyzer.NEURO` object.
 
 - `file_name::String`: name of the file to load
 - `type::Symbol=:eeg`: type of imported data
+    - `:eeg` - EEG
+    - `:meg` - MEG
+    - `:nirs` - fNIRS
+    - `:events` - events
 - `detect_type::Bool=true`: detect channel type based on its label
 
 # Returns
 
-- `obj::NeuroAnalyzer.NEURO`
+- `obj::NeuroAnalyzer.NEURO` - for EEG, MEG, fNIRS data
+- `markers::DataFrame` - for events
 """
-function import_ft(file_name::String; type::Symbol=:eeg, detect_type::Bool=true)::NeuroAnalyzer.NEURO
+function import_ft(file_name::String; type::Symbol=:eeg, detect_type::Bool=true)::Union{NeuroAnalyzer.NEURO, DataFrame}
 
     _wip()
 
-    _check_var(type, [:eeg, :meg, :nirs], "type")
-    @assert type === :eeg "Currently only EEG datatype is supported."
+    _check_var(type, [:eeg, :meg, :nirs, :events], "type")
     file_type = "FT"
     data_type = String(type)
 
@@ -31,191 +35,221 @@ function import_ft(file_name::String; type::Symbol=:eeg, detect_type::Bool=true)
     _info("Reading object: $(string.(keys(dataset))[1])")
     dataset = dataset[string.(keys(dataset))[1]]
 
-    @assert "cfg" in keys(dataset) "Dataset does not contain cfg field."
-    cfg = dataset["cfg"]
-    @assert "hdr" in keys(dataset) "Dataset does not contain hdr field."
-    hdr = dataset["hdr"]
-    @assert "nChans" in keys(hdr) "Dataset header does not contain nChans field."
-    ch_n = Int64(hdr["nChans"])
-    @assert "Fs" in keys(hdr) "Dataset header does not contain Fs field."
-    sampling_rate = round(Int64, hdr["Fs"])
+    if data_type == "events"
 
-    # get channel labels, types and units
-    if length(hdr["label"][:]) > 0 && length(vec(hdr["label"])) == ch_n
-        clabels = string.(hdr["label"][:])
+        @assert "duration" in keys(dataset) "Dataset does not contain duration field."
+        id = dataset["type"][:]
+        duration = dataset["duration"][:]
+        offset = dataset["offset"][:]
+        start = Int64.(dataset["sample"][:])
+        description = dataset["value"][:]
+
+        # find and remove empty records
+        description_idx = findall(x -> length(x) == 0, description)
+        for idx in length(description_idx):-1:1
+            deleteat!(id, idx)
+            deleteat!(duration, idx)
+            deleteat!(offset, idx)
+            deleteat!(start, idx)
+            deleteat!(description, idx)
+        end
+        markers = DataFrame(:id=>string.(id),
+                            :start=>Float64.(start),
+                            :length=>Float64.(duration),
+                            :description=>string.(description),
+                            :channel=>zeros(Int64, length(id)))
+        _info("Imported: $(nrow(markers)) events; events start and length are in samples, use `markers_s2t()` to convert to seconds")
+
+        return markers
+
     else
-        clabels = String[]
-        for idx in 1:ch_n
-            push!(clabels, "ch_$idx")
-        end
-    end
-    clabels = _clean_labels(clabels)
+        @assert "cfg" in keys(dataset) "Dataset does not contain cfg field."
+        cfg = dataset["cfg"]
+        @assert "hdr" in keys(dataset) "Dataset does not contain hdr field."
+        hdr = dataset["hdr"]
[…]
