+++ src/io/import_digitrack.jl
@@ -85,7 +85,7 @@ function import_digitrack(file_name::String; detect_type::Bool=true)::NeuroAnaly
     markers = DataFrame(:id=>String[],
                         :start=>Float64[],
                         :length=>Float64[],
-                        :description=>String[],
+                        :value=>String[],
                         :channel=>Int64[])
 
     time_pts = round.(collect(0:1/sampling_rate:size(data, 2) * size(data, 3) / sampling_rate)[1:end-1], digits=3)
