+++ src/io/import_ft.jl
@@ -20,7 +20,7 @@ Load FieldTrip file (.mat) and return `NeuroAnalyzer.NEURO` object.
 - `obj::NeuroAnalyzer.NEURO` - for EEG, MEG, fNIRS data
 - `markers::DataFrame` - for events
 """
-function import_ft(file_name::String; type::Symbol, detect_type::Bool=true)::Union{NeuroAnalyzer.NEURO, DataFrame}
+function import_ft(file_name::String; type::Symbol, detect_type::Bool=false)::Union{NeuroAnalyzer.NEURO, DataFrame}
 
     _wip()
 
@@ -82,14 +82,14 @@ function import_ft(file_name::String; type::Symbol, detect_type::Bool=true)::Uni
         cfg = dataset["cfg"]
         @assert "hdr" in keys(dataset) "Dataset does not contain hdr field."
         hdr = dataset["hdr"]
+        @assert "fsample" in keys(dataset) "Dataset does not contain fsample field."
+        sampling_rate = round(Int64, dataset["fsample"])
         @assert "nChans" in keys(hdr) "Dataset header does not contain nChans field."
         ch_n = Int64(hdr["nChans"])
-        @assert "Fs" in keys(hdr) "Dataset header does not contain Fs field."
-        sampling_rate = round(Int64, hdr["Fs"])
 
         # get channel labels, types and units
         if length(hdr["label"][:]) > 0 && length(vec(hdr["label"])) == ch_n
-            clabels = string.(hdr["label"][:])
+            clabels = string.(strip.(string.(hdr["label"])))[:]
         else
             clabels = String[]
             for idx in 1:ch_n
@@ -99,9 +99,9 @@ function import_ft(file_name::String; type::Symbol, detect_type::Bool=true)::Uni
         clabels = _clean_labels(clabels)
 
         if detect_type
-            ch_type = _set_channel_types(clabels, "eeg")
+            ch_type = NeuroAnalyzer._set_channel_types(clabels)
             if "chanunit" in keys(hdr)
-                units = string.(hdr["chanunit"][:])
+                units = strip.(string.(hdr["chanunit"][:]))
                 units = replace.(units, "uV"=>"μV")
             else
                 units = [_ch_units(ch_type[idx]) for idx in 1:ch_n]
@@ -110,7 +110,7 @@ function import_ft(file_name::String; type::Symbol, detect_type::Bool=true)::Uni
             if "chantype" in keys(hdr)
                 ch_type = string.(hdr["chantype"][:])
             else
-                ch_type = repeat(["eeg"], ch_n)
+                ch_type = repeat([data_type], ch_n)
             end
             if "chanunit" in keys(hdr)
                 units = string.(hdr["chanunit"][:])
@@ -130,9 +130,15 @@ function import_ft(file_name::String; type::Symbol, detect_type::Bool=true)::Uni
 
         if "time" in keys(dataset)
             if ep_n == 1
-                epoch_time = round.(dataset["time"][1][:], digits=3)
+                epoch_time = dataset["time"][1][:]
+                sign(epoch_time[1]) == 1.0 && (epoch_time .-= (epoch_time[1]))
                 time_pts = dataset["time"][1][:]
-                time_pts .+= abs(time_pts[1])
+                if sign(time_pts[1]) == 1.0
+                    time_pts .-= (time_pts[1])
+                else
+                    time_pts .+= abs(time_pts[1])
+                end
+                epoch_time = round.(epoch_time, digits=3)
                 time_pts = round.(time_pts, digits=3)
             else
                 epoch_time = round.(dataset["time"][1][:], digits=3)
[…]
