+++ src/plots/plot_topo.jl
@@ -162,8 +162,8 @@ function plot_topo(s::Vector{<:Real}; locs::DataFrame, ch::Union{Int64, Vector{I
                            colorbar=cb,
                            palette=pal,
                            aspect_ratio=1,
-                           size=head12 ? size(head_shape) .+ 105 : size(head_shape) .+ 115,
-                           right_margin=-30*Plots.px,
+                           size=head12 ? size(head_shape) .+ 105 : size(head_shape) .+ 110,
+                           right_margin=-10*Plots.px,
                            bottom_margin=10*Plots.px,
                            top_margin=20*Plots.px,
                            left_margin=-30*Plots.px,
