### 2024-11-23 62fdf4d
update: import_ft(), plot_locs_nirs()
