+++ src/io/import_ft.jl
@@ -13,7 +13,7 @@ Load FieldTrip file (.mat) and return `NeuroAnalyzer.NEURO` object.
     - `:meg`: MEG
     - `:nirs`: fNIRS
     - `:events`: events
-- `detect_type::Bool=true`: detect channel type based on its label
+- `detect_type::Bool=false`: detect channel type based on its label
 
 # Returns
 
@@ -27,6 +27,7 @@ function import_ft(file_name::String; type::Symbol, detect_type::Bool=false)::Un
     _check_var(type, [:eeg, :meg, :nirs, :events], "type")
     file_type = "FT"
     data_type = String(type)
+    locs = nothing
 
     @assert isfile(file_name) "File $file_name cannot be loaded."
     dataset = matread(file_name)
@@ -99,7 +100,7 @@ function import_ft(file_name::String; type::Symbol, detect_type::Bool=false)::Un
         clabels = _clean_labels(clabels)
 
         if detect_type
-            ch_type = NeuroAnalyzer._set_channel_types(clabels)
+            ch_type = _set_channel_types(clabels)
             if "chanunit" in keys(hdr)
                 units = strip.(string.(hdr["chanunit"][:]))
                 units = replace.(units, "uV"=>"μV")
@@ -119,6 +120,8 @@ function import_ft(file_name::String; type::Symbol, detect_type::Bool=false)::Un
                 units = repeat(["μV"], ch_n)
             end
         end
+        ch_type = replace.(ch_type, "nirs"=>"nirs_od")
+        ch_type = replace.(ch_type, "stimulus"=>"mrk")
 
         @assert "trial" in keys(dataset) "Dataset does not contain trial field."
         ep_n = "trial" in keys(dataset) ? size(dataset["trial"], 2) : 1
@@ -185,9 +188,9 @@ function import_ft(file_name::String; type::Symbol, detect_type::Bool=false)::Un
                                       channel_order=_sort_channels(ch_type),
                                       reference=ref,
                                       clabels=clabels,
-                                      transducers="Transducer" in keys(hdr["orig"]) ? strip.(string.(hdr["orig"]["Transducer"])) : repeat([""], ch_n),
+                                      transducers="Transducer" in keys(hdr["orig"]) ? string.(strip.(hdr["orig"]["Transducer"])) : repeat([""], ch_n),
                                       units=units,
-                                      prefiltering="PreFilt" in keys(hdr["orig"]) ? strip.(string.(hdr["orig"]["PreFilt"])) : repeat([""], ch_n),
+                                      prefiltering="PreFilt" in keys(hdr["orig"]) ? string.(strip.(hdr["orig"]["PreFilt"])) : repeat([""], ch_n),
                                       line_frequency=50,
                                       sampling_rate=sampling_rate,
                                       gain=ones(ch_n),
@@ -235,6 +238,7 @@ function import_ft(file_name::String; type::Symbol, detect_type::Bool=false)::Un
                     units[ch_idx] = "μV"
                 end
             end
+
             lp = "lowpass" in keys(hdr["orig"]) ? string(round(hdr["orig"]["lowpass"][1], digits=1)) : "?"
             hp = "highpass" in keys(hdr["orig"]) ? string(round(hdr["orig"]["highpass"][1], digits=1)) : "?"
             r = _create_recording_meg(data_type=data_type,
@@ -260,26 +264,72 @@ function import_ft(file_name::String; type::Symbol, detect_type::Bool=false)::Un
 
         elseif data_type == "nirs"
 
+
+            clabels = replace.(clabels, ".0"=>"")
+            clabels = replace.(clabels, " ["=>" ")
+            clabels = replace.(clabels, "nm]"=>"")
+
[…]
