+++ src/io/import_nirs.jl
@@ -191,7 +191,7 @@ function import_nirs(file_name::String)::NeuroAnalyzer.NEURO
                                det_labels=det_labels,
                                opt_labels=opt_labels,
                                sampling_rate=sampling_rate,
-                              bad_channels=zeros(Bool, size(data, 1), 1))
+                               bad_channels=zeros(Bool, size(data, 1), 1))
     e = _create_experiment(name="", notes="", design="")
 
     hdr = _create_header(s,
