### 2024-11-24 aefc91d
update: NeuroAnalyzer.jl [CI SKIP]
