+++ src/NeuroAnalyzer.jl
@@ -221,88 +221,88 @@ end
 # load sub-modules
 
 # internal functions
+include("internal/channels.jl")
 include("internal/check.jl")
 include("internal/components.jl")
 include("internal/create_header.jl")
 include("internal/draw_head.jl")
+include("internal/epochs.jl")
 include("internal/fiff.jl")
 include("internal/fir_response.jl")
+include("internal/gdf_etp.jl")
 include("internal/gpu.jl")
+include("internal/gradient.jl")
+include("internal/gui.jl")
 include("internal/interpolate.jl")
 include("internal/labeled_matrix.jl")
 include("internal/labels.jl")
 include("internal/len.jl")
 include("internal/locs.jl")
-include("internal/gradient.jl")
-include("internal/gui.jl")
-include("internal/epochs.jl")
-include("internal/channels.jl")
 include("internal/markers.jl")
 include("internal/misc.jl")
 include("internal/ml.jl")
 include("internal/plots.jl")
+include("internal/recorder.jl")
 include("internal/reflect_chop.jl")
+include("internal/rt_plotter.jl")
 include("internal/select.jl")
+include("internal/statistics.jl")
+include("internal/tester.jl")
 include("internal/time.jl")
 include("internal/wl2ext.jl")
-include("internal/gdf_etp.jl")
-include("internal/statistics.jl")
-include("internal/recorder.jl")
-include("internal/rt_plotter.jl")
-include("internal/tester.jl")
 # analyze
-include("analyze/ampdiff.jl")
+include("analyze/acor.jl")
+include("analyze/acov.jl")
 include("analyze/amp.jl")
-include("analyze/rmse.jl")
-include("analyze/snr.jl")
-include("analyze/erp_peaks.jl")
+include("analyze/ampdiff.jl")
+include("analyze/axc2frq.jl")
 include("analyze/band_asymmetry.jl")
 include("analyze/band_mpower.jl")
 include("analyze/band_power.jl")
 include("analyze/coherence.jl")
-include("analyze/cpsd.jl")
-include("analyze/erop.jl")
-include("analyze/eros.jl")
-include("analyze/frqinst.jl")
-include("analyze/hjorth.jl")
-include("analyze/peak_frq.jl")
-include("analyze/psd.jl")
-include("analyze/psd_rel.jl")
-include("analyze/psd_slope.jl")
-include("analyze/sef.jl")
-include("analyze/spectrogram.jl")
-include("analyze/spectrum.jl")
-include("analyze/total_power.jl")
-include("analyze/mep_peaks.jl")
 include("analyze/corm.jl")
 include("analyze/covm.jl")
+include("analyze/cph.jl")
+include("analyze/cpsd.jl")
 include("analyze/dissimilarity.jl")
 include("analyze/entropy.jl")
 include("analyze/envelopes.jl")
+include("analyze/erop.jl")
+include("analyze/eros.jl")
+include("analyze/erp_peaks.jl")
+include("analyze/frqinst.jl")
 include("analyze/ged.jl")
+include("analyze/hjorth.jl")
 include("analyze/hrv.jl")
-include("analyze/mi.jl")
-include("analyze/stationarity.jl")
-include("analyze/tkeo.jl")
-include("analyze/cph.jl")
 include("analyze/ispc.jl")
 include("analyze/itpc.jl")
+include("analyze/mep_peaks.jl")
+include("analyze/mi.jl")
+include("analyze/pacor.jl")
+include("analyze/peak_frq.jl")
 include("analyze/phdiff.jl")
 include("analyze/phsd.jl")
 include("analyze/pli.jl")
-include("analyze/acor.jl")
-include("analyze/acov.jl")
[…]
