### 2024-11-25 eee550c
update: NeuroAnalyzer.jl
