+++ src/utils/na.jl
@@ -1,449 +1,0 @@
-export na_info
-export na_plugins_reload
-export na_plugins_list
-export na_plugins_remove
-export na_plugins_install
-export na_plugins_update
-export na_set_use_cuda
-export na_set_progress_bar
-export na_set_prefs
-export na_set_verbose
-export na_version
-
-# suppress console output
-_log_off() = Logging.disable_logging(Logging.Warn)
-# restore console output
-_log_on() = Logging.disable_logging(Logging.Debug)
-
-_info(s::String) = verbose && @info s
-_warn(s::String) = verbose && @warn s
-_deprecated(s::String) = verbose && @error "Function $s() is deprecated."
-_deprecated(s1::String, s2::String) = verbose && @error "Function $s1() is deprecated, please use $s2() instead."
-_wip() = allow_wip ? (@warn "This function has the WIP (Work In Progress) status and is not ready for production use.") : (@error "This function has the WIP (Work In Progress) status and is not ready for production use.")
-
-"""
-    na_info()
-
-Show NeuroAnalyzer and imported packages versions.
-
-# Arguments
-
-Nothing
-
-# Returns
-
-Nothing
-"""
-function na_info()::Nothing
-
-    println("     NeuroAnalyzer: $(NeuroAnalyzer.VER)")
-    println("NeuroAnalyzer path: $(NeuroAnalyzer.PATH)")
-    println("     Julia version: $VERSION")
-    if CUDA.functional()
-        println("              CUDA: $(CUDA.runtime_version()) (use_cuda = $use_cuda)")
-    else
-        println("              CUDA: not available (use_cuda = $use_cuda)")
-    end
-    println("      Plugins path: $plugins_path")
-    println("    Resources path: $res_path")
-    println(" Show progress bar: $progress_bar")
-    println("           Verbose: $verbose")
-    println("           Threads: $(Threads.nthreads()) [set using `JULIA_NUM_THREADS` environment variable or Julia --threads command-line option]")
-    Threads.nthreads() < length(Sys.cpu_info()) || @info "For best performance, `JULIA_NUM_THREADS` ($(Threads.nthreads())) should be less than number of CPU threads ($(length(Sys.cpu_info())))."
-    if "JULIA_COPY_STACKS" in keys(ENV) && ENV["JULIA_COPY_STACKS"] == "1"
-        @info "Environment variable `JULIA_COPY_STACKS` is set to 1, multi-threading may not work correctly"
-    end
-    println()
-
-    na_pkg = dirname(dirname(Base.find_package("NeuroAnalyzer")))
-
-    if isfile(joinpath(na_pkg, "Manifest.toml"))
-        println("Imported packages:")
-        required_packages = [
-            "Cairo",
-            "ColorSchemes",
-            "ContinuousWavelets",
-            "CSV",
-            "CubicSplines",
-            "CUDA",
-            "DataFrames",
-            "Deconvolution",
-            "DICOM",
-            "Distances",
-            "DSP",
-            "FFTW",
-            "FileIO",
-            "FindPeaks1D",
-            "FIRLSFilterDesign",
-            "FourierTools",
-            "GeometryBasics",
-            "Git",
-            "GLM",
-            "Gtk",
-            "HypothesisTests",
-            "Images",
-            "ImageBinarization",
-            "ImageMorphology",
-            "InformationMeasures",
-            "Interpolations",
-            "Jacobi",
-            "JLD2",
-            "JSON",
-            "LibSerialPort",
-            "LinRegOutliers",
[…]
