### 2024-11-26 dd66e8a
add: generate_ssp_projectors(), apply_ssp_projectors() update: import_ft()
