### 2024-11-26 651fbfc
add: apply_ssp_projectors!()
