+++ src/process/ssp.jl
@@ -10,7 +10,7 @@ Generate SSP projectors from embedded projections.
 # Arguments
 
 - `obj::NeuroAnalyzer.NEURO`
-- `projectors::Union{Int64, Vector{Int64}}=0`: list of projections used for generating projectors
+- `proj::Union{Int64, Vector{Int64}}=0`: list of projections used for generating projectors, by default use all available projections
 
 # Returns
 
@@ -18,24 +18,25 @@ Named tuple containing:
 - `ssp_projectors::Matrix{Float64}` : projectors
 - `U::Matrix{Float64}}`: SVD U orthogonal matrix
 """
-function generate_ssp_projectors(obj::NeuroAnalyzer.NEURO; projectors::Union{Int64, Vector{Int64}}=0)::@NamedTuple{ssp_projectors::Matrix{Float64}, U::Matrix{Float64}}
+function generate_ssp_projectors(obj::NeuroAnalyzer.NEURO; proj::Union{Int64, Vector{Int64}}=0)::@NamedTuple{ssp_projectors::Matrix{Float64}, U::Matrix{Float64}}
 
-    @assert "ssp_data" in keys(obj.header.recording) "OBJ does not contain SSP projectors."
-    @assert size(obj.header.recording[:ssp_data], 1) > 0 "OBJ does not contain SSP projectors."
+    @assert "ssp_data" in keys(obj.header.recording) "OBJ does not contain SSP projections."
+    @assert size(obj.header.recording[:ssp_data], 1) > 0 "OBJ does not contain SSP projections."
 
-    if projectors == 0
-        projectors = 1:size(obj.header.recording[:ssp_data], 1)
+    # by default use all available projections
+    if proj == 0
+        proj = 1:size(obj.header.recording[:ssp_data], 1)
     end
 
-    if isa(projectors, Int64)
-        @assert projectors >= 1 && projectors <= size(obj.header.recording[:ssp_data], 1) "projectors must be in [1, $(size(obj.header.recording[:ssp_data], 1))]."
+    if isa(proj, Int64)
+        @assert proj >= 1 && proj <= size(obj.header.recording[:ssp_data], 1) "proj must be in [1, $(size(obj.header.recording[:ssp_data], 1))]."
     else
-        projectors = sort(projectors)
-        @assert projectors[1] >= 1 && projectors[end] <= size(obj.header.recording[:ssp_data], 1) "projectors must be in [1, $(size(obj.header.recording[:ssp_data], 1))]."
+        proj = sort(proj)
+        @assert proj[1] >= 1 && proj[end] <= size(obj.header.recording[:ssp_data], 1) "proj must be in [1, $(size(obj.header.recording[:ssp_data], 1))]."
     end
 
-    # extract projectors
-    ssp_projectors = obj.header.recording[:ssp_data][projectors, :]'
+    # extract projections
+    ssp_projectors = obj.header.recording[:ssp_data][proj, :]'
 
     # reorthogonalize the vectors
     U, S, _ = svd(ssp_projectors)
@@ -60,25 +61,25 @@ Apply SSP projectors from embedded projections.
 # Arguments
 
 - `obj::NeuroAnalyzer.NEURO`
-- `projectors::Union{Int64, Vector{Int64}}=0`: list of projections used for generating projectors
+- `proj::Union{Int64, Vector{Int64}}=0`: list of projections used for generating projectors, by default use all available projections
 
 # Returns
 
 - `obj_new::NeuroAnalyzer.NEURO`
 """
[…]
