+++ src/NeuroAnalyzer.jl
@@ -9,7 +9,7 @@ module NeuroAnalyzer
 
 # set constants
 
-const VER = v"0.24.12-dev"
+const VER = v"0.24.12"
 const allow_wip = occursin("dev", string(VER))          # false for the stable branch, true for the devel branch
 const io = PipeBuffer()                                 # required for interactive preview
 const data_types = ["eeg",
