+++ Artifacts.toml
@@ -6,8 +6,8 @@ git-tree-sha1 = "6f696f162cb48a93809736f4311ee729e58f4ee4"
     sha256 = "9ba60c01280e01ed1adc4c2d7c2c143a5eac57257b6dba10e3ad33a0884db336"
 
 [NeuroAnalyzer_test-files]
-git-tree-sha1 = "0206d559b1fe92f5f611e4cbf19761e8799f49a4"
+git-tree-sha1 = "ad21c4aff4d1435644796802e2b950741a3052f7"
 
     [[NeuroAnalyzer_test-files.download]]
     url = "https://neuroanalyzer.org/artifacts/neuroanalyzer-test-files.tar.gz"
-    sha256 = "d9108b400047d89d38cd353a2109e9560a7411f952a0650d2517625079f68e69"
+    sha256 = "ef81d54a1d7e64b6761c9d07cd43ce63eedcdefc264ce02f96cc344e232ce00c"
