### 2024-11-27 8c440c6
update: version 0.24.12
