+++ src/NeuroAnalyzer.jl
@@ -218,18 +218,287 @@ end
 
 # load sub-modules
 
-_load_functions("internal")
-_load_functions("utils")
-_load_functions("io")
-_load_functions("locs")
-_load_functions("edit")
-_load_functions("process")
-_load_functions("analyze")
-_load_functions("plots")
-_load_functions("gui")
-_load_functions("statistics")
-_load_functions("study")
-_load_functions("recorder")
-_load_functions("stim")
+# internal
+include("internal/channels.jl")
+include("internal/check.jl")
+include("internal/components.jl")
+include("internal/create_header.jl")
+include("internal/draw_head.jl")
+include("internal/epochs.jl")
+include("internal/fiff.jl")
+include("internal/fir_response.jl")
+include("internal/gdf_etp.jl")
+include("internal/gpu.jl")
+include("internal/gradient.jl")
+include("internal/gui.jl")
+include("internal/interpolate.jl")
+include("internal/labeled_matrix.jl")
+include("internal/labels.jl")
+include("internal/len.jl")
+include("internal/locs.jl")
+include("internal/map_channels.jl")
+include("internal/markers.jl")
+include("internal/misc.jl")
+include("internal/ml.jl")
+include("internal/plots.jl")
+include("internal/recorder.jl")
+include("internal/reflect_chop.jl")
+include("internal/rt_plotter.jl")
+include("internal/select.jl")
+include("internal/statistics.jl")
+include("internal/tester.jl")
+include("internal/time.jl")
+include("internal/wl2ext.jl")
+
+# utils
+include("utils/apply.jl")
+include("utils/array.jl")
+include("utils/components.jl")
+include("utils/fft.jl")
+include("utils/findpeaks.jl")
+include("utils/frequency.jl")
+include("utils/generate.jl")
+include("utils/gradient.jl")
+include("utils/info.jl")
+include("utils/labels.jl")
+include("utils/make_table.jl")
+include("utils/matrix.jl")
+include("utils/misc.jl")
+include("utils/note.jl")
+include("utils/pad.jl")
+include("utils/phase.jl")
+include("utils/play.jl")
+include("utils/time.jl")
+include("utils/to_df.jl")
+include("utils/vector.jl")
+include("utils/view_header.jl")
+include("utils/wavelets.jl")
+
+# io
+include("io/export_csv.jl")
+include("io/export_locs.jl")
+include("io/export_markers.jl")
+include("io/import_alice4.jl")
+include("io/import_bdf.jl")
+include("io/import_bv.jl")
+include("io/import_cnt.jl")
+include("io/import_csv.jl")
+include("io/import_dat.jl")
+include("io/import_digitrack.jl")
+include("io/import_duomag.jl")
+include("io/import_edf_annotations.jl")
+include("io/import_edf.jl")
+include("io/import_fiff.jl")
+include("io/import_ft.jl")
+include("io/import_gdf.jl")
+include("io/import.jl")
+include("io/import_locs.jl")
+include("io/import_montage.jl")
+include("io/import_ncs.jl")
+include("io/import_nirs.jl")
+include("io/import_nirx.jl")
+include("io/import_npy.jl")
+include("io/import_nwb.jl")
+include("io/import_set.jl")
+include("io/import_snirf.jl")
+include("io/import_thymatron.jl")
+include("io/import_xdf.jl")
+include("io/load_locs.jl")
+include("io/save_load.jl")
+
+# locs
+include("locs/add_locs.jl")
+include("locs/center.jl")
+include("locs/convert.jl")
+include("locs/details.jl")
[…]
