+++ .woodpecker.yml
@@ -6,8 +6,6 @@ steps:
       - event: [push, manual]   # workflow should run if you execute it manually and on each push
     secrets: [ cbtoken, cbmail, ftp, ftp_pass, ftp_user ]
     commands:
-      - uname -a
-      - julia --version
       - export CI="1"
 
       - mkdir ~/NeuroAnalyzer
@@ -43,6 +41,7 @@ steps:
     when:
       - event: manual   # workflow should run if you execute it manually
     commands:
+      - export CI="1"
       - mkdir ~/NeuroAnalyzer
       - mkdir ~/NeuroAnalyzer/plugins
       - apt-get update
