+++ Artifacts.toml
@@ -2,12 +2,12 @@
 git-tree-sha1 = "6f696f162cb48a93809736f4311ee729e58f4ee4"
 
     [[NeuroAnalyzer_resources.download]]
-    url = "https://neuroanalyzer.org/artifacts/neuroanalyzer-resources.tar.gz"
+    url = "https://neuroanalyzer.org/artifacts/neuroanalyzer-resources-dev.tar.gz"
     sha256 = "9ba60c01280e01ed1adc4c2d7c2c143a5eac57257b6dba10e3ad33a0884db336"
 
 [NeuroAnalyzer_test-files]
 git-tree-sha1 = "ad21c4aff4d1435644796802e2b950741a3052f7"
 
     [[NeuroAnalyzer_test-files.download]]
-    url = "https://neuroanalyzer.org/artifacts/neuroanalyzer-test-files.tar.gz"
+    url = "https://neuroanalyzer.org/artifacts/neuroanalyzer-test-files-dev.tar.gz"
     sha256 = "ef81d54a1d7e64b6761c9d07cd43ce63eedcdefc264ce02f96cc344e232ce00c"
