### 2024-11-28 b9b3981
update: plot_locs()
