### 2024-11-28 5bd32b8
add: lat_idx() update: plot_locs(), import_ft()
