+++ src/NeuroAnalyzer.jl
@@ -408,6 +408,7 @@ include("analyze/hjorth.jl")
 include("analyze/hrv.jl")
 include("analyze/ispc.jl")
 include("analyze/itpc.jl")
+include("analyze/lat_idx.jl")
 include("analyze/mep_peaks.jl")
 include("analyze/mi.jl")
 include("analyze/pacor.jl")
