+++ test/analyze.jl
@@ -920,4 +920,8 @@ ph, f = phsd(e10, ch="all")
 @test symmetry(v) == 5
 @test symmetry(e10, ch="Fp1") == [1.001563721657545 0.9527078565980168 1.0285261489698891 0.9452887537993921 1.0 0.970746728252502 0.9219219219219219 1.0496397117694156 1.0173364854215918 0.9393939393939394]
 
+@info "Test: lat_idx()"
+@test lat_idx(e10) isa Float64
+@test lat_idx(e10, frq=(1, 3.5)) isa Float64
+
 true
