### 2024-11-28 514ce02
update: lat_idx()
