### 2024-11-28 149b5e5
update: lat_idx()
