### 2024-11-29 97f0115
update: get_channel()
