+++ src/analyze/acov.jl
@@ -140,7 +140,7 @@ Calculate autocovariance. For ERP return trial-averaged autocovariance.
 # Arguments
 
 - `obj::NeuroAnalyzer.NEURO`
-- `ch::Union{String, Vector{String}}`: channel name or list of channel names
+- `ch::Union{String, Vector{String}, Regex}`: channel name or list of channel names
 - `l::Int64=1`: lags range is `0:lag` [samples]
 - `demean::Bool=true`: demean signal before computing autocovariance
 - `biased::Bool=true`: calculate biased or unbiased autocovariance
@@ -155,7 +155,7 @@ Named tuple containing:
 - `ac::Array{Float64, 3}`
 - `l::Vector{Float64}`: lags [s]
 """
-function acov(obj::NeuroAnalyzer.NEURO; ch::Union{String, Vector{String}}, l::Real=1, demean::Bool=true, biased::Bool=true, method::Symbol=:sum)::@NamedTuple{ac::Array{Float64, 3}, l::Vector{Float64}}
+function acov(obj::NeuroAnalyzer.NEURO; ch::Union{String, Vector{String}, Regex}, l::Real=1, demean::Bool=true, biased::Bool=true, method::Symbol=:sum)::@NamedTuple{ac::Array{Float64, 3}, l::Vector{Float64}}
 
     @assert l <= size(obj, 2) "l must be ≤ $(size(obj, 2))."
     @assert l >= 0 "l must be ≥ 0."

[…]
