### 2024-12-04 b2ea86e
fix: cmp_test()
