+++ src/recorder/tpt.jl
@@ -1,14 +1,16 @@
-export ieda
-export eda
+export itpt
+export tpt
 
 """
-    ieda(; <keyword arguments>)
+    itpt(; <keyword arguments>)
 
-Record electrodermal activity (EDA), also called Galvanic Skin Response (GSR) or skin conductance, in GUI mode. EDA is recorded using Groove GSR sensor via Arduino attached to the PC via USB cable (virtual serial port). Sampling rate is 50 Hz.
+Record Two-point Pinch Test (TPT) in GUI mode. TPT is recorded using MMA7660 accelerometer via Arduino attached to the PC via USB cable (virtual serial port). Sampling rate is 50 Hz.
 
 # Arguments
 
-- `duration::Int64=20`: recording duration in seconds
+- `duration::Int64=20`: single trial duration in seconds
+- `trials::Int64=2`: number of trials
+- `interval::Int64=2`: interval between trials in seconds
 - `port_name::String="/dev/ttyUSB0"`: serial port to which the Arduino is connected
 
 # Returns
@@ -16,7 +18,7 @@ Record electrodermal activity (EDA), also called Galvanic Skin Response (GSR) or
 - `obj_new::NeuroAnalyzer.NEURO`
 - `Nothing` if recording fails
 """
-function ieda(; duration::Int64=20, port_name::String="/dev/ttyUSB0")::Union{NeuroAnalyzer.NEURO, Nothing}
+function itpt(; duration::Int64=20, trials::Int64=2, interval::Int64=2, port_name::String="/dev/ttyUSB0")::Union{NeuroAnalyzer.NEURO, Nothing}
 
     sp = _serial_open(port_name, baudrate=19200)
     @assert !isnothing(sp) @info "Serial port $port_name is not available"
@@ -24,7 +26,12 @@ function ieda(; duration::Int64=20, port_name::String="/dev/ttyUSB0")::Union{Neu
     # sampling rate is 50 Hz = 20 ms per loop
     fs = 50
     t = collect(0:1/fs:duration)
-    eda_signal = repeat([NaN], length(t))
+    tpt_ch_x = repeat([NaN], length(t))
+    tpt_ch_y = repeat([NaN], length(t))
+    tpt_ch_z = repeat([NaN], length(t))
+    tpt_ch_accx = repeat([NaN], length(t))
+    tpt_ch_accy = repeat([NaN], length(t))
+    tpt_ch_accz = repeat([NaN], length(t))
 
     p = Plots.plot(ylims=(0, 10),
                    xlims=(t[1], t[end]),
@@ -37,7 +44,7 @@ function ieda(; duration::Int64=20, port_name::String="/dev/ttyUSB0")::Union{Neu
                    xtickfontsize=8,
                    ytickfontsize=8)
 
-    win = GtkWindow("NeuroRecorder: ieda()", p.attr[:size][1], p.attr[:size][2] + 40)
+    win = GtkWindow("NeuroRecorder: itpt()", p.attr[:size][1], p.attr[:size][2] + 40)
     set_gtk_property!(win, :border_width, 20)
     set_gtk_property!(win, :resizable, false)
     set_gtk_property!(win, :has_resize_grip, false)
@@ -69,12 +76,10 @@ function ieda(; duration::Int64=20, port_name::String="/dev/ttyUSB0")::Union{Neu
 
     @guarded draw(can) do widget
         p = Plots.plot(t,
-                       eda_signal,
-                       mc=:black,
+                       tpt_ch_x,
                        ms=0.5,
                        lw=0.5,
-                       lc=:black,
-                       ylims=(0, 10),
+                       ylims=(0, 20),
                        xlims=(t[1], t[end]),
                        legend=false,
[…]
