+++ src/recorder/tpt.jl
@@ -1,24 +1,21 @@
-export itpt
-export tpt
+export itptr
+export tptr
 
 """
-    itpt(; <keyword arguments>)
+    itptr(; <keyword arguments>)
 
 Record Two-point Pinch Test (TPT) in GUI mode. TPT is recorded using MMA7660 accelerometer via Arduino attached to the PC via USB cable (virtual serial port). Sampling rate is 50 Hz.
 
 # Arguments
 
-- `duration::Int64=20`: single trial duration in seconds
-- `trials::Int64=2`: number of trials
-- `interval::Int64=2`: interval between trials in seconds
+- `duration::Int64=20`: recording duration in seconds
 - `port_name::String="/dev/ttyUSB0"`: serial port to which the Arduino is connected
 
 # Returns
 
 - `obj_new::NeuroAnalyzer.NEURO`
-- `Nothing` if recording fails
 """
-function itpt(; duration::Int64=20, trials::Int64=2, interval::Int64=2, port_name::String="/dev/ttyUSB0")::Union{NeuroAnalyzer.NEURO, Nothing}
+function itptr(; duration::Int64=20, port_name::String="/dev/ttyUSB0")::NeuroAnalyzer.NEURO
 
     sp = _serial_open(port_name, baudrate=19200)
     @assert !isnothing(sp) @info "Serial port $port_name is not available"
@@ -44,7 +41,7 @@ function itpt(; duration::Int64=20, trials::Int64=2, interval::Int64=2, port_nam
                    xtickfontsize=8,
                    ytickfontsize=8)
 
-    win = GtkWindow("NeuroRecorder: itpt()", p.attr[:size][1], p.attr[:size][2] + 40)
+    win = GtkWindow("NeuroRecorder: itptr()", p.attr[:size][1], p.attr[:size][2] + 40)
     set_gtk_property!(win, :border_width, 20)
     set_gtk_property!(win, :resizable, false)
     set_gtk_property!(win, :has_resize_grip, false)
@@ -160,30 +157,28 @@ function itpt(; duration::Int64=20, trials::Int64=2, interval::Int64=2, port_nam
     @async Gtk.gtk_main()
     wait(cnd)
 
-    if length(tpt_ch_x) > 0
-        tpt_ch_x = tpt_ch_x[1:(end - 1)]
-        tpt_ch_y = tpt_ch_y[1:(end - 1)]
-        tpt_ch_z = tpt_ch_z[1:(end - 1)]
-        tpt_ch_accx = tpt_ch_accx[1:(end - 1)]
-        tpt_ch_accy = tpt_ch_accy[1:(end - 1)]
-        tpt_ch_accz = tpt_ch_accz[1:(end - 1)]
-        t = round.(t[1:(end - 1)], digits=3)
-        tpt_signal = Matrix([tpt_ch_x tpt_ch_y tpt_ch_z tpt_ch_accx tpt_ch_accy tpt_ch_accz]')
-        tpt_signal = reshape(tpt_signal, 6, :, 1)
-        obj = create_object(data_type="tpt")
-        add_channel!(obj, data=tpt_signal, label=["pos_x", "pos_y", "pos_z", "acc_x", "acc_y", "acc_z"], type=["orient", "orient", "orient", "accel", "accel", "accel"], unit=["", "", "", "m/s²", "m/s²", "m/s²"])
-        create_time!(obj, fs=fs)
-        return obj
-    else
-        return nothing
-    end
+    tpt_ch_x = tpt_ch_x[1:(end - 1)]
+    tpt_ch_y = tpt_ch_y[1:(end - 1)]
+    tpt_ch_z = tpt_ch_z[1:(end - 1)]
[…]
