### 2024-12-05 7cc4159
fix: _ticks()
