+++ src/internal/plots.jl
@@ -29,15 +29,19 @@ function _ylims(s::AbstractVector)::Tuple{Real, Real}
 end
 
 function _ticks(t::Union{AbstractVector, AbstractRange})::AbstractVector
-    if length(t) >= 3
-        if t[2] - t[1] == t[3] - t[2]
-            tc = linspace(round(t[1]), round(t[end]), 11)
-        else
-            tc = collect(floor(t[1], digits=2):((ceil(t[end]) - floor(t[1])) / 10):ceil(t[end], digits=2))
-        end
+#    if length(t) >= 3
+#        if t[2] - t[1] == t[3] - t[2]
+#            tc = linspace(round(t[1]), round(t[end]), length(round.(Int64, t[1]:t[end])))
+#        else
+#            tc = collect(floor(t[1], digits=2):((ceil(t[end]) - floor(t[1])) / 10):ceil(t[end], digits=2))
+#        end
+#    end
+    if t[end] == round(Int64, t[end])
+        tc = linspace(round(t[1]), round(t[end]), length(round.(Int64, t[1]:((t[end] - t[1])/10):t[end])))
+    else
+        tc = linspace(t[1], t[end], length(round.(t[1]:((t[end] - t[1])/10):t[end], digits=2)))
     end
-    tc[end] = t[end]
-    tc = round.(tc)
+    tc = round.(tc, digits=2)
     return tc
 end
 
