### 2024-12-05 36c2335
fix: hrv_detect() add: tpt_detect() add: tpt_analyze()
