+++ src/analyze/hrv.jl
@@ -21,7 +21,7 @@ function hrv_detect(obj::NeuroAnalyzer.NEURO)::@NamedTuple{nn_seg::Vector{Float6
     @assert "ecg" in obj.header.recording[:channel_type] "OBJ does not contain ECG channel."
     ch = get_channel(obj, type="ecg")
     _info("ECG channel found: $(labels(obj)[ch])")
-    ecg = eeg.data[ch, :, :][:]
+    ecg = obj.data[ch, :, :][:]
     r_idx, _ = findpeaks1d(ecg, height=mean(ecg) + 2*std(ecg))
 
     # convert to ms
