+++ src/NeuroAnalyzer.jl
@@ -427,6 +427,7 @@ include("analyze/spectrum.jl")
 include("analyze/stationarity.jl")
 include("analyze/symmetry.jl")
 include("analyze/tkeo.jl")
+include("analyze/tpt.jl")
 include("analyze/total_power.jl")
 include("analyze/xcor.jl")
 include("analyze/xcov.jl")
