### 2024-12-05 40c929b
add: ftt_analyze()
