+++ src/NeuroAnalyzer.jl
@@ -428,6 +428,7 @@ include("analyze/stationarity.jl")
 include("analyze/symmetry.jl")
 include("analyze/tkeo.jl")
 include("analyze/tpt.jl")
+include("analyze/ftt.jl")
 include("analyze/total_power.jl")
 include("analyze/xcor.jl")
 include("analyze/xcov.jl")
