+++ src/tester/tpt_old.jl
@@ -1,0 +1,294 @@
+export itpt
+export tpt
+
+"""
+    itpt(; <keyword arguments>)
+
+Perform Two-point Pinch Test (TPT) in GUI mode. TPT is recorded using MMA7660 accelerometer via Arduino attached to the PC via USB cable (virtual serial port). Sampling rate is 50 Hz.
+
+# Arguments
+
+- `duration::Int64=20`: recording duration in seconds
+- `trials::Int64=2`: number of trials
+- `interval::Int64=2`: interval between trials in seconds
+- `port_name::String="/dev/ttyUSB0"`: serial port to which the Arduino is connected
+
+# Returns
+
+- `obj_new::NeuroAnalyzer.NEURO`
+"""
+function itpt(; duration::Int64=20, trials::Int64=2, interval::Int64=2, port_name::String="/dev/ttyUSB0")::NeuroAnalyzer.NEURO
+
+    sp = _serial_open(port_name, baudrate=19200)
+    @assert !isnothing(sp) @info "Serial port $port_name is not available"
+
+    img1 = read_from_png(joinpath(res_path, "finger_nopinch.png"))
+    img2 = read_from_png(joinpath(res_path, "finger_pinch.png"))
+
+    win = GtkWindow("NeuroRecorder: itpt()", img1.width, img1.height + 100)
+    set_gtk_property!(win, :border_width, 20)
+    set_gtk_property!(win, :resizable, false)
+    set_gtk_property!(win, :has_resize_grip, false)
+    set_gtk_property!(win, :window_position, 3)
+    set_gtk_property!(win, :startup_id, "org.neuroanalyzer")
+
+    can = GtkCanvas(img1.width, img1.height)
+
+    g1 = GtkGrid()
+    set_gtk_property!(g1, :column_homogeneous, true)
+    set_gtk_property!(g1, :column_spacing, 20)
+    set_gtk_property!(g1, :row_spacing, 20)
+    g2 = GtkGrid()
+    set_gtk_property!(g2, :column_homogeneous, true)
+    set_gtk_property!(g2, :column_spacing, 20)
+    set_gtk_property!(g2, :row_spacing, 20)
+
+    bt_start = GtkButton("START")
+    set_gtk_property!(bt_start, :tooltip_text, "Start the test")
+
+    lb_status1 = GtkLabel("Status:")
+    lb_status2 = GtkLabel("READY TO START")
+    set_gtk_property!(lb_status1, :halign, 2)
+    set_gtk_property!(lb_status2, :halign, 1)
+    lb_trial1 = GtkLabel("Trial #:")
+    lb_trial2 = GtkLabel("-")
+    set_gtk_property!(lb_trial1, :halign, 2)
+    set_gtk_property!(lb_trial2, :halign, 1)
+    lb_interval1 = GtkLabel("Interval #:")
+    lb_interval2 = GtkLabel("-")
+    set_gtk_property!(lb_interval1, :halign, 2)
+    set_gtk_property!(lb_interval2, :halign, 1)
+
+    g1[1:3, 1] = can
+    g1[1, 2] = lb_status1
+    g1[3, 2] = lb_status2
+    g1[1, 3] = lb_trial1
+    g1[3, 3] = lb_trial2
+    g1[1, 4] = lb_interval1
+    g1[3, 4] = lb_interval2
+    g1[1:3, 5] = GtkLabel("")
+    g1[1:3, 6] = bt_start
+    vbox1 = GtkBox(:v)
+    push!(vbox1, g1)
+
+    push!(win, vbox1)
+    showall(win)
+
+    @guarded draw(can) do widget
+        ctx = getgc(can)
[…]
