+++ src/tester/tpt.jl
@@ -1,10 +1,10 @@
-export itptr
-export tptr
+export itpt
+export tpt
 
 """
-    itptr(; <keyword arguments>)
+    itpt(; <keyword arguments>)
 
-Record Two-point Pinch Test (TPT) in GUI mode. TPT is recorded using MMA7660 accelerometer via Arduino attached to the PC via USB cable (virtual serial port). Sampling rate is 50 Hz.
+Perform Two-point Pinch Test (TPT) in GUI mode. TPT is recorded using MMA7660 accelerometer via Arduino attached to the PC via USB cable (virtual serial port). Sampling rate is 50 Hz.
 
 # Arguments
 
@@ -15,39 +15,31 @@ Record Two-point Pinch Test (TPT) in GUI mode. TPT is recorded using MMA7660 acc
 
 - `obj_new::NeuroAnalyzer.NEURO`
 """
-function itptr(; duration::Int64=20, port_name::String="/dev/ttyUSB0")::NeuroAnalyzer.NEURO
+function itpt(; duration::Int64=20, port_name::String="/dev/ttyUSB0")::NeuroAnalyzer.NEURO
 
     sp = _serial_open(port_name, baudrate=19200)
     @assert !isnothing(sp) @info "Serial port $port_name is not available"
 
+    img1 = read_from_png(joinpath(res_path, "finger_nopinch.png"))
+    img2 = read_from_png(joinpath(res_path, "finger_pinch.png"))
+
     # sampling rate is 50 Hz = 20 ms per loop
     fs = 50
     t = collect(0:1/fs:duration)
-    tpt_ch_x = repeat([NaN], length(t))
-    tpt_ch_y = repeat([NaN], length(t))
-    tpt_ch_z = repeat([NaN], length(t))
-    tpt_ch_accx = repeat([NaN], length(t))
-    tpt_ch_accy = repeat([NaN], length(t))
-    tpt_ch_accz = repeat([NaN], length(t))
+    tpt_ch_x = zeros(length(t))
+    tpt_ch_y = zeros(length(t))
+    tpt_ch_z = zeros(length(t))
+    tpt_ch_accx = zeros(length(t))
+    tpt_ch_accy = zeros(length(t))
+    tpt_ch_accz = zeros(length(t))
 
-    p = Plots.plot(ylims=(0, 10),
-                   xlims=(t[1], t[end]),
-                   legend=false,
-                   palette=:darktest,
-                   size=(800, 400),
-                   margins=20Plots.px,
-                   xlabelfontsize=8,
-                   ylabelfontsize=8,
-                   xtickfontsize=8,
-                   ytickfontsize=8)
-
-    win = GtkWindow("NeuroRecorder: itptr()", p.attr[:size][1], p.attr[:size][2] + 40)
+    win = GtkWindow("NeuroRecorder: itpt()", img1.width, img1.height + 100)
     set_gtk_property!(win, :border_width, 20)
     set_gtk_property!(win, :resizable, false)
     set_gtk_property!(win, :has_resize_grip, false)
     set_gtk_property!(win, :window_position, 3)
     set_gtk_property!(win, :startup_id, "org.neuroanalyzer")
-    can = GtkCanvas(Int32(p.attr[:size][1]), Int32(p.attr[:size][2]))
+    can = GtkCanvas(img1.width, img1.height)
     g = GtkGrid()
     set_gtk_property!(g, :column_homogeneous, false)
     set_gtk_property!(g, :column_spacing, 5)
@@ -72,34 +64,8 @@ function itptr(; duration::Int64=20, port_name::String="/dev/ttyUSB0")::NeuroAna
     showall(win)
 
     @guarded draw(can) do widget
-        p = Plots.plot(t,
-                       tpt_ch_x,
-                       ms=0.5,
-                       lw=0.5,
[…]
