+++ src/NeuroAnalyzer.jl
@@ -241,6 +241,7 @@ include("internal/map_channels.jl")
 include("internal/markers.jl")
 include("internal/misc.jl")
 include("internal/ml.jl")
+include("internal/peaks.jl")
 include("internal/plots.jl")
 include("internal/recorder.jl")
 include("internal/reflect_chop.jl")
