### 2025-01-11 #7465 @Kevin-Mattheus-Moerman
@TheCedarPrince, @vladdez, @ferchaure, @behinger, @AJQuinn, @azraq27, @rob-luke, @kinleyid, @paulmthompson, @alexmorley  would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software](https://joss.theoj.org)? The work is titled `NeuroAnalyzer: Julia toolbox for analyzing neurophysiological data` and features the use of the Julia programming language for "importing, editing, processing, visualizing, and analyzing EEG, MEP and EDA data".

JOSS reviews take place on GitHub and focus on [the software :floppy_disk: ](https://github.com/JuliaHealth/NeuroAnalyzer.jl) as well as [a short paper :scroll: ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07465/joss.07465/10.21105.joss.07465.pdf). We typically aim for a 2-3 week review but we can be more flexible if needed. You can find more about our review process in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).  Here is an example of an ongoing review if you want to see what it is like: https://github.com/openjournals/joss-reviews/issues/7564.

If you are interested you can let me know by responding here.

Thanks!

@TheCedarPrince I see you are also a contributor on the project, please do confirm if you have a conflict of interest or not. Thanks.
