### 2025-01-11 #7465 @AdamWysokinski
There is a minor type in line 133 (missing space) of the paper -- may I correct it now?
