### 2025-01-14 ec1d644
update: cmp_test()
