+++ src/statistics/cmp_test.jl
@@ -3,7 +3,7 @@ export cmp_test
 """
     cmp_test(seg1, seg2, paired, alpha, type, exact)
 
-Compare two vectors; Kruskall-Wallis test is used first, next t-test (paired on non-paired) or non-parametric test (paired: Wilcoxon signed rank, non-paired: Mann-Whitney U test) is applied.
+Compare two vectors. Jarque–Bera is used first to test normality of distribution, next variances are tested using F-test. Finally, paired or non-paired t-test or non-parametric test (Wilcoxon signed rank for paired data; Mann-Whitney U test for non-paired data) is applied. Permutation-based testing is also 
 
 # Arguments
 
@@ -11,35 +11,43 @@ Compare two vectors; Kruskall-Wallis test is used first, next t-test (paired on 
 - `s2::AbstractVector`
 - `paired::Bool`
 - `alpha::Float64=0.05`: confidence level
-- `type::Symbol=:auto`: choose test automatically (`:auto`), permutation-based (`:perm`), parametric (`:p`) or non-parametric (`:np`)
+- `type::Symbol=:auto`:
+    - `:auto`: choose test automatically
+    - `:perm`: permutation-based
+    - `:p`: parametric
+    - `:np`: non-parametric
 - `exact::Bool=false`: if true, use exact Wilcoxon test
 - `nperm::Int64=1000`: number of permutation for `:perm` method
 
 # Returns
 
-Named tuple containing for type !== `:perm`:
-- `t`: test results
-- `ts::Tuple{Float64, String}`: test statistics
-- `tc::Tuple{Float64, Float64}`: test statistics confidence interval
-- `df::Float64`: degrees of freedom
-- `p::Float64`: p-value
-
-Named tuple containing for type === `:perm`:
+For permutation-based test:
 - `t::Tuple{perm_diff::Vector{Float64}, obs_diff::Float64}`: test results: (permutation difference, observed difference)
 - `p1::Float64`: one-sided p-value
 - `p2::Float64`: two-sided p-value
+
+Otherwise, named tuple containing:
+- `t`: statistical test
+- `ts::Tuple{Float64, String}`: test statistic value and name
+- `tc::Tuple{Float64, Float64}`: test statistic confidence interval
+- `df::Float64`: degrees of freedom
+- `p::Float64`: p-value
+
 """
-function cmp_test(s1::AbstractVector, s2::AbstractVector; paired::Bool, alpha::Float64=0.05, type::Symbol=:auto, exact::Bool=false, nperm::Int64=1000)::Union{@NamedTuple{t::OneSampleTTest, ts::Tuple{Float64, String}, tc::Tuple{Float64, Float64}, df::Float64, p::Float64}, @NamedTuple{t::EqualVarianceTTest, ts::Tuple{Float64, String}, tc::Tuple{Float64, Float64}, df::Float64, p::Float64}, @NamedTuple{t::UnequalVarianceTTest, ts::Tuple{Float64, String}, tc::Tuple{Float64, Float64}, df::Float64, p::Float64}, @NamedTuple{t::ApproximateMannWhitneyUTest{Float64}, ts::Tuple{Float64, String}, tc::Float64, df::Float64, p::Float64}, @NamedTuple{t::@NamedTuple{perm_diff::Vector{Float64}, obs_diff::Float64}, p1::Float64, p2::Float64}}
[…]
