+++ test/internal.jl
@@ -154,5 +154,7 @@ t, et = NeuroAnalyzer._get_t(e10)
 @test NeuroAnalyzer._set_defaults("a", "b", "c", "d", "e", "f") == ("a", "b", "c")
 @test NeuroAnalyzer._set_defaults("default", "default", "default", "d", "e", "f") == ("d", "e", "f")
 @test NeuroAnalyzer._midxy(1, 1, 4, 4) == (2.5, 2.5)
+@test NeuroAnalyzer._between(1, 1, 2.0) == true
+@test NeuroAnalyzer._between(0.9, 1.0, 2) == false
 
 true
