+++ src/gui/ifilter.jl
@@ -1,0 +1,235 @@
+export ifilter
+
+"""
+    ifilter(obj)
+
+Interactive filter design.
+
+# Arguments
+
+- `obj::NeuroAnalyzer.NEURO`
+
+# Returns
+
+- `obj_new::NeuroAnalyzer.NEURO`
+"""
+function ifilter(obj::NeuroAnalyzer.NEURO)::Nothing
+
+    fprototypes = [:butterworth, :chebyshev1, :chebyshev2, :elliptic, :fir, :firls, :iirnotch, :remez]
+    fprototype = fprototypes[1]
+    ftypes = [:lp, :hp, :bs, :bp]
+    ftype = ftypes[1]
+    cutoff = 0.01
+    order = 8
+    rp = 2
+    rs = 20
+    bw = 2
+    w = nothing
+
+    fs = sr(obj)
+
+    p = plot_filter_response(fs=fs, n=epoch_len(obj), fprototype=fprototype, ftype=ftype, cutoff=cutoff, order=order, rp=rp, rs=rs, bw=bw, w=w)
+
+    win = GtkWindow("NeuroAnalyzer: ifilter()", 1300, 800)
+    set_gtk_property!(win, :border_width, 5)
+    set_gtk_property!(win, :resizable, true)
+    set_gtk_property!(win, :has_resize_grip, false)
+    set_gtk_property!(win, :window_position, 3)
+    set_gtk_property!(win, :startup_id, "org.neuroanalyzer")
+    can = GtkCanvas(Int32(p.attr[:size][1]), Int32(p.attr[:size][2]))
+
+    g = GtkGrid()
+    g_opts = GtkGrid()
+    set_gtk_property!(g, :column_homogeneous, false)
+    set_gtk_property!(g_opts, :column_homogeneous, false)
+    set_gtk_property!(g, :column_spacing, 5)
+    set_gtk_property!(g, :row_spacing, 5)
+    set_gtk_property!(g_opts, :row_spacing, 5)
+    set_gtk_property!(g_opts, :column_spacing, 5)
+
+    bt_close = GtkButton("Close")
+    set_gtk_property!(bt_close, :tooltip_text, "Close this window")
+
+    cb_mono = GtkCheckButton()
+    set_gtk_property!(cb_mono, :tooltip_text, "Use color or gray palette")
+
+    combo_fprototype = GtkComboBoxText()
+    for idx in uppercase.(String.(fprototypes))
+        push!(combo_fprototype, idx)
+    end
+    set_gtk_property!(combo_fprototype, :active, 0)
+    set_gtk_property!(combo_fprototype, :tooltip_text, "Filter prototype")
+
+    combo_ftype = GtkComboBoxText()
+    for idx in uppercase.(String.(ftypes))
+        push!(combo_ftype, idx)
+    end
+    set_gtk_property!(combo_ftype, :active, 0)
+    set_gtk_property!(combo_ftype, :tooltip_text, "Filter type")
+
+    entry_cutoff1 = GtkSpinButton(0.1, fs / 2 - 0.1, 0.1)
+    set_gtk_property!(entry_cutoff1, :value, cutoff)
+    set_gtk_property!(entry_cutoff1, :tooltip_text, "Filter cutoff in Hz (lower)")
+
+    entry_cutoff2 = GtkSpinButton(0.2, fs / 2, 0.1)
+    set_gtk_property!(entry_cutoff2, :value, cutoff)
+    set_gtk_property!(entry_cutoff2, :sensitive, false)
+    set_gtk_property!(entry_cutoff2, :tooltip_text, "Filter cutoff in Hz (upper)")
+
+    entry_order = GtkSpinButton(2, 256, 2)
+    set_gtk_property!(entry_order, :value, order)
[…]
