### 2025-01-20 e8ab48b
update: ifilter() fix: keyboard shortcuts in GUI
