+++ src/gui/ifilter.jl
@@ -11,15 +11,15 @@ Interactive filter design.
 
 # Returns
 
-- `obj_new::NeuroAnalyzer.NEURO`
+- `flt::Union{Nothing, Vector{Float64}, ZeroPoleGain{:z, ComplexF64, ComplexF64, Float64}, Biquad{:z, Float64}}`
 """
-function ifilter(obj::NeuroAnalyzer.NEURO)::Nothing
+function ifilter(obj::NeuroAnalyzer.NEURO)::Union{Nothing, Vector{Float64}, ZeroPoleGain{:z, ComplexF64, ComplexF64, Float64}, Biquad{:z, Float64}}
 
     fprototypes = [:butterworth, :chebyshev1, :chebyshev2, :elliptic, :fir, :firls, :iirnotch, :remez]
     fprototype = fprototypes[1]
     ftypes = [:lp, :hp, :bs, :bp]
     ftype = ftypes[1]
-    cutoff = 0.1
+    cutoff = 1
     order = 8
     rp = 2
     rs = 20
@@ -79,7 +79,7 @@ function ifilter(obj::NeuroAnalyzer.NEURO)::Nothing
     set_gtk_property!(entry_cutoff2, :sensitive, false)
     set_gtk_property!(entry_cutoff2, :tooltip_text, "Filter cutoff in Hz (upper)")
 
-    entry_order = GtkSpinButton(2, 256, 2)
+    entry_order = GtkSpinButton(1, 256, 1)
     set_gtk_property!(entry_order, :value, order)
     set_gtk_property!(entry_order, :tooltip_text, "Filter order (6 dB/octave) for IIR filters\nNumber of taps for REMEZ filter\nAttenuation (× 4 dB) for FIR filter")
 
@@ -144,42 +144,51 @@ function ifilter(obj::NeuroAnalyzer.NEURO)::Nothing
     showall(win)
 
     @guarded draw(can) do widget
-        if mod(get_gtk_property(entry_order, :value, Int64), 2) != 0
-            warn_dialog("Order must be even!")
-            Gtk.@sigatom begin
-                set_gtk_property!(entry_order, :value, order)
-            end
-        else
-            ctx = getgc(can)
-            mono = get_gtk_property(cb_mono, :active, Bool)
-            cutoff1 = get_gtk_property(entry_cutoff1, :value, Float64)
-            cutoff2 = get_gtk_property(entry_cutoff2, :value, Float64)
-            order = get_gtk_property(entry_order, :value, Int64)
-            order = get_gtk_property(entry_order, :value, Int64)
-            rp = get_gtk_property(entry_rp, :value, Float64)
-            rs = get_gtk_property(entry_rs, :value, Float64)
-            bw = get_gtk_property(entry_bw, :value, Float64)
-            fprototype = fprototypes[get_gtk_property(combo_fprototype, :active, Int64) + 1]
-            ftype = ftypes[get_gtk_property(combo_ftype, :active, Int64) + 1]
-            fprototype === :iirnotch && (ftype = :bs)
-            if ftype === :bp || ftype === :bs
-                if fprototype !== :iirnotch
-                    cutoff = (cutoff1, cutoff2)
-                else
-                    cutoff = cutoff1
-                end
+        ctx = getgc(can)
+        mono = get_gtk_property(cb_mono, :active, Bool)
+        cutoff1 = get_gtk_property(entry_cutoff1, :value, Float64)
+        cutoff2 = get_gtk_property(entry_cutoff2, :value, Float64)
+        order = get_gtk_property(entry_order, :value, Int64)
+        rp = get_gtk_property(entry_rp, :value, Float64)
+        rs = get_gtk_property(entry_rs, :value, Float64)
+        bw = get_gtk_property(entry_bw, :value, Float64)
[…]
