+++ src/process/filter.jl
@@ -1,4 +1,5 @@
 export filter_apply
+export filter_apply!
 export filter_create
 export filter
 export filter!
@@ -27,7 +28,7 @@ Create IIR or FIR filter.
 - `cutoff::Union{Real, Tuple{Real, Real}}=0`: filter cutoff in Hz (must be a pair of frequencies for `:bp` and `:bs`)
 - `n::Int64`: signal length in samples
 - `fs::Int64`: sampling rate
-- `order::Int64=8`: filter order (6 dB/octave), number of taps for `:remez`, attenuation (× 4 dB) for `:fir` filters
+- `order::Int64=4`: filter order (6 dB/octave) for `:butterworth`, `:chebyshev1`, `:chebyshev2`, `:elliptic` (for these filters is multiplied by two since must be even), number of taps for `:remez`, attenuation (× 4 dB) for `:fir` filters
 - `rp::Real=-1`: ripple amplitude in dB in the pass band; default: 0.0025 dB for `:elliptic`, 2 dB for others
 - `rs::Real=-1`: ripple amplitude in dB in the stop band; default: 40 dB for `:elliptic`, 20 dB for others
 - `bw::Real=-1`: bandwidth for `:iirnotch` and :remez filters
@@ -37,7 +38,7 @@ Create IIR or FIR filter.
 
 - `flt::Union{Vector{Float64}, ZeroPoleGain{:z, ComplexF64, ComplexF64, Float64}, Biquad{:z, Float64}}`
 """
-function filter_create(; fprototype::Symbol, ftype::Union{Nothing, Symbol}=nothing, cutoff::Union{Real, Tuple{Real, Real}}=0, n::Int64, fs::Int64, order::Int64=8, rp::Real=-1, rs::Real=-1, bw::Real=-1, w::Union{Nothing, AbstractVector, <:Real}=nothing)::Union{Vector{Float64}, ZeroPoleGain{:z, ComplexF64, ComplexF64, Float64}, Biquad{:z, Float64}}
+function filter_create(; fprototype::Symbol, ftype::Union{Nothing, Symbol}=nothing, cutoff::Union{Real, Tuple{Real, Real}}=0, n::Int64, fs::Int64, order::Int64=4, rp::Real=-1, rs::Real=-1, bw::Real=-1, w::Union{Nothing, AbstractVector, <:Real}=nothing)::Union{Vector{Float64}, ZeroPoleGain{:z, ComplexF64, ComplexF64, Float64}, Biquad{:z, Float64}}
 
     _check_var(fprototype, [:butterworth, :chebyshev1, :chebyshev2, :elliptic, :fir, :firls, :iirnotch, :remez], "fprototype")
     if fprototype !== :iirnotch
@@ -46,11 +47,11 @@ function filter_create(; fprototype::Symbol, ftype::Union{Nothing, Symbol}=nothi
     end
 
     @assert fs >= 1 "fs must be ≥ 1."
-    fprototype !== :iirnotch && @assert order > 1 "order must be > 1."
+
+    !(fprototype in [:iirnotch, :remez, :fir]) && @assert order > 1 "order must be > 1."
     @assert order <= n "order must be ≤ signal length ($n)."
-    if !(fprototype in [:iirnotch, :remez, :fir])
-        @assert order >= 2 "order must be ≥ 2."
-        @assert iseven(order) "order must be even."
+    if fprototype in [:butterworth, :chebyshev1, :chebyshev2, :elliptic]
+        order *= 2
     end
 
     if isa(w, AbstractVector)
@@ -339,6 +340,92 @@ function filter_apply(s::AbstractVector; flt::Union{Vector{Float64}, ZeroPoleGai
 end
 
 """
[…]

[…]
