### 2025-01-22 a391d5b
update: cmp_test()
