### 2025-01-26 #7465 @TheCedarPrince
@Kevin-Mattheus-Moerman :

> @TheCedarPrince I see you are also a contributor on the project, please do confirm if you have a conflict of interest or not. Thanks.

To disclose, I am the leader of the JuliaHealth organization which @AdamWysokinski is a member of. I have used NeuroAnalyzer.jl before for some of my own work which resulted in the following publication: https://www.frontiersin.org/journals/network-physiology/articles/10.3389/fnetp.2024.1478280/full . In the course of this, I contributed a small bug fix in the process of using the tool.

If that surmounts to a sufficient conflict of interest, then please let me know and I will withdraw my name as reviewer. Otherwise, happy to be a reviewer here as well. :-)

Cheers,

~ tcp 🌳
