### 2025-01-31 #7734 @ferchaure
Another small detail, on https://neuroanalyzer.org/tut-download.html

it should be:

```
using Pkg
Pkg.add("NeuroAnalyzer")
```
