### 2025-01-31 #7734 @AdamWysokinski
> Another small detail, on https://neuroanalyzer.org/tut-download.html

Well spotted, thanks. Fixed now.
