+++ Project.toml
@@ -80,7 +80,7 @@ ColorSchemes = "3"
 ContinuousWavelets = "1"
 CubicSplines = "0.2"
 DICOM = "0.11"
-DSP = "0.7"
+DSP = "0.7, 0.8"
 DataFrames = "1.6"
 Deconvolution = "1"
 Distances = "0.10"
@@ -119,7 +119,7 @@ Polynomials = "3.2, 4"
 Preferences = "1.4"
 PrettyTables = "2.3"
 ProgressMeter = "1.10"
-SavitzkyGolay = "0.7, 0.8"
+SavitzkyGolay = "0.7, 0.8, 0.9"
 ScatteredInterpolation = "0.3"
 Simpson = "1.0.1"
 Static = "1.1"
