### 2025-02-10 #7734 @TheCedarPrince
Hi @AdamWysokinski and @Kevin-Mattheus-Moerman ,

Finished my review, some notes for Adam:

- Typo in line 156: Dataseris -> Datseris
- In your "statement of need" section, could you explicitly say who your target audience is? Is it neuroscientists who need a highly performant neuroscience toolbox? I think you have the answer in your summary already but it would be great to have it in this section more explicitly.
- In line 30, you make a few performance claims -- do you have a citation to back those up anywhere?
- In line 36, do you have a reference for this statement?
- In line 42, could you be more explicit about what you mean here with Julia being beautifully designed? How did that inform your motivation in using Julia?
- Line 129: "is stored using standard Julia Array's and can be easily exported as a DataFrame"

Once those fixes and adjustments are made, this paper has my approval. Great work Adam!
