+++ joss/paper.bib
@@ -96,4 +96,28 @@
   urldate      = {2021-10-18},
 }
 
-@Comment{jabref-meta: databaseType:bibtex;}
+@Article{bezanson_2014,
+  author       = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+  title        = {Julia: {A} Fresh Approach to Numerical Computing},
+  journal      = {CoRR},
+  volume       = {abs/1411.1607},
+  year         = {2014},
+  url          = {http://arxiv.org/abs/1411.1607},
+  eprinttype    = {arXiv},
+  eprint       = {1411.1607},
+  timestamp    = {Mon, 13 Aug 2018 16:47:09 +0200},
+  biburl       = {https://dblp.org/rec/journals/corr/BezansonEKS14.bib},
+  bibsource    = {dblp computer science bibliography, https://dblp.org}
+}
+
+@article{pal_2024,
+  title = {A next-generation dynamic programming language Julia: Its features and applications in biological science},
+  journal = {Journal of Advanced Research},
+  volume = {64},
+  pages = {143-154},
+  year = {2024},
+  issn = {2090-1232},
+  doi = {https://doi.org/10.1016/j.jare.2023.11.015},
+  url = {https://www.sciencedirect.com/science/article/pii/S2090123223003521},
+  author = {Pal, Soumen and Bhattacharya, Manojit and Dash, Snehasish and Lee, Sang-Soo and Chakraborty, Chiranjib},
+}
