### 2025-02-10 a8f123e
update: JOSS paper
