+++ joss/paper.md
@@ -37,7 +37,7 @@ NeuroAnalyzer is a Julia [@bezanson_2017] toolbox for analyzing neurophysiologic
 
 There are many excellent MATLAB and Python based EEG/MEG/NIRS applications (e.g. EEGLAB [@delorme_2011], Fieldtrip [@oostenveld_2011], Brainstorm or MNE [@gramfort_2013]). They have been in development for many years and are well established in the scientific community. Many state-of-the-art papers were published using data prepared using these programs. However, compared with Python and MATLAB, there are many advantages of Julia, which underlie my decision to start developing such a toolbox in Julia. I believe that Julia is the future of scientific computing and scientific data analysis [@selvaraj_2022]. Major advantages of Julia are listed in Julia documentation:
 
-- Julia is fast [@bezanson_2018]. In many situations Julia is considerably faster than Python (without having to use numba/cython) and MATLAB (see https://benchmarksgame-team.pages.debian.net/benchmarksgame/index.html for example benchmarks). Moreover, Julia provides unlimited scalability. Julia programs can easily be ran on a large cluster or across distributed computers.
+- Julia is fast [@bezanson_2018]. In many situations Julia is considerably faster than Python (without having to use numba/cython) and MATLAB (see [Benchmarks Game](https://benchmarksgame-team.pages.debian.net/benchmarksgame/index.html) for example benchmarks). Moreover, Julia provides unlimited scalability. Julia programs can easily be ran on a large cluster or across distributed computers.
 - Julia is open-source and free. Increasing MATLAB licensing costs are prohibitive to individual researchers and many research institutions.
 - From its very beginning Julia is being focused on scientific computations [@bezanson_2014]. Currently only Julia, C, C++ and Fortran belong to the HPC (High Performance Computing) Petaflop Club. Julia is designed for distributed and parallel computations, making it great for distributed analyzes of large data sets.
 - Most of the Julia packages are written in pure Julia. It’s easier to understand and modify their code if you already know Julia.
