### 2025-02-12 b4f024e
add: effsize_p1g()
