### 2025-02-12 cb005f7
update: effsize_p2g()
