+++ src/analyze/mdiff.jl
@@ -1,0 +1,45 @@
+export mdiff
+
+"""
+    mdiff(obj1, obj2; <keyword arguments>)
+
+Calculate mean difference and 95% confidence interval for two channels.
+
+# Arguments
+
+- `obj1::NeuroAnalyzer.NEURO`
+- `obj2:NeuroAnalyzer.NEURO`
+- `ch1::Union{String, Vector{String}}`: list of channels
+- `ch2::Union{String, Vector{String}}`: list of channels
+- `ep1::Union{Int64, Vector{Int64}, AbstractRange}=_c(nepochs(obj1))`: default use all epochs
+- `ep2::Union{Int64, Vector{Int64}, AbstractRange}=_c(nepochs(obj2))`: default use all epochs
+- `n::Int64`: number of bootstraps
+- `method::Symbol[:absdiff, :diff2int]`
+    - `:absdiff`: maximum difference
+    - `:diff2int`: integrated area of the squared difference
+
+# Returns
+
+Named tuple containing:
+- `st::Matrix{Float64}`
+- `sts::Vector{Float64}`
+- `p::Vector{Float64}`
+"""
+function mdiff(obj1::NeuroAnalyzer.NEURO, obj2::NeuroAnalyzer.NEURO; ch1::Union{String, Vector{String}}, ch2::Union{String, Vector{String}}, ep1::Union{Int64, Vector{Int64}, AbstractRange}=_c(nepochs(obj1)), ep2::Union{Int64, Vector{Int64}, AbstractRange}=_c(nepochs(obj2)), n::Int64=3, method::Symbol=:absdiff)::@NamedTuple{st::Matrix{Float64}, sts::Vector{Float64}, p::Vector{Float64}}
+
+    ch1 = get_channel(obj1, ch=ch1)
+    ch2 = get_channel(obj2, ch=ch2)
+    @assert length(ch1) == length(ch2) "ch1 and ch2 must have the same length."
+
+    _check_epochs(obj1, ep1)
+    _check_epochs(obj2, ep2)
+    @assert length(ep1) == length(ep2) "ep1 and ep2 must have the same length."
+    @assert epoch_len(obj1) == epoch_len(obj2) "OBJ1 and OBJ2 must have the same epoch lengths."
+
+    length(ep1) == 1 && (ep1 = [ep1])
+    length(ep2) == 1 && (ep2 = [ep2])
+
+    st, sts, p = @views NeuroStats.mdiff(obj1.data[ch1, :, ep1], obj2.data[ch2, :, ep2], n=n, method=method)
+
+    return (st=st, sts=sts, p=p)
+end
