### 2025-02-27 6e01423
update: significant speed increase due to @threads :static
