+++ src/analyze/entropy.jl
@@ -61,7 +61,7 @@ function entropy(s::AbstractArray)::@NamedTuple{ent::Matrix{Float64}, sent::Matr
     leent = zeros(ch_n, ep_n)
 
     @inbounds for ep_idx in 1:ep_n
-        Threads.@threads for ch_idx in 1:ch_n
+        Threads.@threads :static for ch_idx in 1:ch_n
             ent[ch_idx, ep_idx], sent[ch_idx, ep_idx], leent[ch_idx, ep_idx] = @views entropy(s[ch_idx, :, ep_idx])
         end
     end
@@ -141,7 +141,7 @@ function negentropy(s::AbstractArray)::Matrix{Float64}
     ne = zeros(ch_n, ep_n)
 
     @inbounds for ep_idx in 1:ep_n
-        Threads.@threads for ch_idx in 1:ch_n
+        Threads.@threads :static for ch_idx in 1:ch_n
             ne[ch_idx, ep_idx] = @views negentropy(s[ch_idx, :, ep_idx])
         end
     end
