+++ src/process/filter_mavg.jl
@@ -66,7 +66,7 @@ function filter_mavg(s::AbstractArray; k::Int64=8, t::Real=0, window::AbstractVe
     s_filtered = similar(s)
 
     @inbounds for ep_idx in 1:ep_n
-        Threads.@threads :static for ch_idx in 1:ch_n
+        Threads.@threads :greedy for ch_idx in 1:ch_n
             s_filtered[ch_idx, :, ep_idx] = @views filter_mavg(s[ch_idx, :, ep_idx], k=k, t=t, window=window)
         end
     end
