+++ src/process/filter_poly.jl
@@ -93,7 +93,7 @@ function filter_poly(s::AbstractArray; order::Int64=8, window::Int64=10)::Array{
     s_filtered = similar(s)
 
     @inbounds for ep_idx in 1:ep_n
-        Threads.@threads :static for ch_idx in 1:ch_n
+        Threads.@threads :greedy for ch_idx in 1:ch_n
             s_filtered[ch_idx, :, ep_idx] = @views filter_poly(s[ch_idx, :, ep_idx], order=order, window=window)
         end
     end
