+++ src/edit/reflect_chop.jl
@@ -28,7 +28,7 @@ function reflect(obj::NeuroAnalyzer.NEURO; n::Int64=sr(obj))::NeuroAnalyzer.NEUR
     s = zeros(ch_n, epoch_len(obj) + 2 * n, ep_n)
 
     @inbounds for ep_idx in 1:ep_n
-        Threads.@threads :static for ch_idx in 1:ch_n
+        Threads.@threads :greedy for ch_idx in 1:ch_n
             s1 = obj_new.data[:, 1:n, ep_idx]
             s2 = obj_new.data[:, end:-1:(end - n + 1), ep_idx]
             @views s[ch_idx, :, ep_idx] = _reflect(obj.data[ch_idx, :, ep_idx], s1[ch_idx, :], s2[ch_idx, :])
@@ -96,7 +96,7 @@ function chop(obj::NeuroAnalyzer.NEURO; n::Int64=sr(obj))::NeuroAnalyzer.NEURO
     s = zeros(ch_n, epoch_len(obj) - 2 * n, ep_n)
 
     @inbounds for ep_idx in 1:ep_n
-        Threads.@threads :static for ch_idx in 1:ch_n
+        Threads.@threads :greedy for ch_idx in 1:ch_n
             @views s[ch_idx, :, ep_idx] = _chop(obj.data[ch_idx, :, ep_idx], n)
         end
     end
