+++ src/process/cw.jl
@@ -84,7 +84,7 @@ function cw_trans(s::AbstractArray; wt::T)::Array{Float64, 4} where {T<:CWT}
     ct = zeros(ch_n, l, ep_len, ep_n)
 
     @inbounds for ep_idx in 1:ep_n
-        Threads.@threads :static for ch_idx in 1:ch_n
+        Threads.@threads :greedy for ch_idx in 1:ch_n
             ct[ch_idx, :, :, ep_idx] = @views cw_trans(s[ch_idx, :, ep_idx], wt=wt)
         end
     end
