### 2025-02-28 #7734 @kinleyid
I've completed my review. I opened issues 4-15 but they were addressed before I had a chance to link them here :) Happy to recommend acceptance.
