### 2025-03-09 e0f54fa
fix: average_epochs()
