### 2025-03-10 b7ea598
add: exclude_bads preference
