+++ src/analyze/ispc.jl
@@ -55,7 +55,7 @@ Named tuple containing:
 """
 function ispc(obj::NeuroAnalyzer.NEURO; ch::Union{String, Vector{String}, Regex})::@NamedTuple{ispc_val::Array{Float64, 3}, ispc_ang::Array{Float64, 3}}
 
-    ch = get_channel(obj, ch=ch)
+    ch = exclude_bads ? get_channel(obj, ch=ch, exclude="bad") : get_channel(obj, ch=ch, exclude="")
     ch_n = length(ch)
     ep_n = nepochs(obj)
 
@@ -104,8 +104,8 @@ Named tuple containing:
 """
 function ispc(obj1::NeuroAnalyzer.NEURO, obj2::NeuroAnalyzer.NEURO; ch1::Union{String, Vector{String}}, ch2::Union{String, Vector{String}}, ep1::Union{Int64, Vector{Int64}, AbstractRange}=_c(nepochs(obj1)), ep2::Union{Int64, Vector{Int64}, AbstractRange}=_c(nepochs(obj2)))::@NamedTuple{ispc_val::Matrix{Float64}, ispc_ang::Matrix{Float64}, s_diff::Array{Float64, 3}, ph_diff::Array{Float64, 3}, s1_phase::Array{Float64, 3}, s2_phase::Array{Float64, 3}}
 
-    ch1 = get_channel(obj1, ch=ch1)
-    ch2 = get_channel(obj2, ch=ch2)
+    ch1 = exclude_bads ? get_channel(obj1, ch=ch1, exclude="bad") : get_channel(obj1, ch=ch1, exclude="")
+    ch2 = exclude_bads ? get_channel(obj2, ch=ch2, exclude="bad") : get_channel(obj2, ch=ch2, exclude="")
     @assert length(ch1) == length(ch2) "ch1 and ch2 must have the same length."
 
     _check_epochs(obj1, ep1)
