### 2025-03-11 4c8939c
update: default palette
