### 2025-03-12 71bd3a5
fix: list_component()
