### 2025-03-12 feee02d
update: normalize()
