+++ src/plots/plot_psd.jl
@@ -124,7 +124,7 @@ function plot_psd(sf::Vector{Float64}, sp::Matrix{Float64}; clabels::Vector{Stri
     # each channel is between -1.0 and +1.0
     for idx in 1:ch_n
         # scale by 0.5 so maxima do not overlap
-        sp[idx, :] = @views normalize(sp[idx, :], method=:minmax) .* 0.5 .+ (idx - 1)
+        sp[idx, :] = @views normalize_minmax(sp[idx, :]) .* 0.5 .+ (idx - 1)
     end
 
     if frq === :lin
