+++ src/process/normalize.jl
@@ -1,1104 +1,4 @@
 export normalize
-export normalize_zscore
-export normalize_minmax
-export normalize_n
-export normalize_log
-export normalize_gauss
-export normalize_log10
-export normalize_neglog
-export normalize_neglog10
-export normalize_neg
-export normalize_pos
-export normalize_perc
-export normalize_invroot
-export normalize_softmax
-export normalize_sigmoid
-export normalize_mad
-export normalize_rank
-export normalize_fisher
-
-"""
-    normalize(s, n; <keyword arguments>)
-
-Normalize.
-
-# Arguments
-
-- `s::AbstractVector`
-- `n::Real=1.0`
-- `method::Symbol`:
-    - `:zscore`: by z-score
-    - `:minmax`: in [-1, +1]
-    - `:log`: using log-transformation
-    - `:log10`: using log10-transformation
-    - `:neglog`: using -log-transformation
-    - `:neglog10`: using -log10-transformation
-    - `:neg`: in [-∞, 0]
-    - `:pos`: in [0, +∞]
-    - `:perc`: in percentages
-    - `:gauss`: to Gaussian
-    - `:invroot`: to inverse root: 1/sqrt(x)
-    - `:n`: in [0, n], default is [0, 1]; <keyword arguments>) .+ n1`
-    - `:softmax`: using softmax function: exp(x_i) / sum(exp(x))
-    - `:sigmoid`: using sigmoid function: 1 /  1 + exp(-x_i)
-    - `:mad`: by MAD
-    - `:rank`: using tiedranks
-    - `:none`
-
-# Returns
-
-- `normalized::AbstractVector`
-"""
-function normalize(s::AbstractVector, n::Float64=1.0; method::Symbol)::AbstractVector
-
-    _check_var(method, [:zscore, :minmax, :log, :log10, :neglog, :neglog10, :neg, :pos, :perc, :gauss, :invroot, :n, :softmax, :sigmoid, :mad, :rank, :none], "method")
-
-    if method === :zscore
-        return normalize_zscore(s)
-    elseif method === :minmax
-        return normalize_minmax(s)
-    elseif method === :log
-        return normalize_log(s)
-    elseif method === :log10
-        return normalize_log10(s)
-    elseif method === :neglog
-        return normalize_neglog(s)
-    elseif method === :neglog10
-        return normalize_neglog10(s)
-    elseif method === :neg
-        return normalize_neg(s)
-    elseif method === :pos
-        return normalize_pos(s)
-    elseif method === :perc
-        return normalize_perc(s)
-    elseif method === :gauss
-        return normalize_gauss(s)
-    elseif method === :invroot
-        return normalize_invroot(s)
-    elseif method === :n
-        return normalize_n(s, n)
-    elseif method === :softmax
-        return normalize_softmax(s)
-    elseif method === :sigmoid
-        return normalize_sigmoid(s)
-    elseif method === :mad
-        return normalize_mad(s)
-    elseif method === :rank
-        return normalize_rank(s)
-    elseif method === :none
-        return s
-    end
-
-end
-
-"""
-    normalize(s, n; <keyword arguments>)
-
-Normalize.
-
-# Arguments
-
-- `s::AbstractArray`
-- `n::Real=1.0`
-- `bych::Bool=false`: if true, normalize each channel separately
-- `method::Symbol`:
-    - `:zscore`: by z-score
-    - `:minmax`: in [-1, +1]
-    - `:log`: using log-transformation
-    - `:log10`: using log10-transformation
[…]
