+++ test/process.jl
@@ -245,60 +245,6 @@ e10_int = lrinterpolate_channel(e10_tmp, ch="Fp1", ep=1)
 @test e10_int.data[1, :, 1] != e10_tmp.data[1, :, 1]
 
 @info "Test: normalize()"
-@test NeuroAnalyzer.normalize(v1, method=:none) == v1
-@test NeuroAnalyzer.normalize(m1, method=:none) == m1
-@test NeuroAnalyzer.normalize(a1, method=:none) == a1
-@test normalize_zscore(v1) == [-1.2649110640673518, -0.6324555320336759, 0.0, 0.6324555320336759, 1.2649110640673518]
-@test normalize_zscore(m1, bych=true) == [-1 0 1; -1 0 1]
-@test normalize_zscore(m1, bych=false) == [-1.3363062095621219 -0.8017837257372732 -0.2672612419124244; 0.2672612419124244 0.8017837257372732 1.3363062095621219]
-@test normalize_minmax(v1) == [-1.0, -0.5, 0.0, 0.5, 1.0]
-@test normalize_minmax(m1, bych=true) == [-1 0 1; -1 0 1]
-@test normalize_minmax(m1, bych=false) == [-1.0 -0.6 -0.19999999999999996; 0.19999999999999996 0.6000000000000001 1.0]
-@test normalize_n(v1) == [0.0, 0.25, 0.5, 0.75, 1.0]
-@test normalize_n(m1, bych=true) == [0.0 0.5 1.0; 0.0 0.5 1.0]
-@test normalize_n(m1, bych=false) == [0.0 0.2 0.4; 0.6 0.8 1.0]
-@test normalize_log(v1) == [1.0986122886681098, 1.3862943611198906, 1.6094379124341003, 1.791759469228055, 1.9459101490553132]
-@test normalize_log(m1, bych=true) == [1.0986122886681098 1.3862943611198906 1.6094379124341003; 2.1972245773362196 2.302585092994046 2.3978952727983707]
-@test normalize_log(m1, bych=false) == [1.0986122886681098 1.3862943611198906 1.6094379124341003; 1.791759469228055 1.9459101490553132 2.0794415416798357]
-@test normalize_gauss(v1) == [-0.8047189562170504, -0.3465735902799727, 0.0, 0.3465735902799726, 0.8047189562170504]
-@test normalize_gauss(m1, bych=true) == [-0.5493061443340549 0.0 0.5493061443340549; -0.5493061443340549 0.0 0.5493061443340549]
-@test normalize_gauss(m1, bych=false) == [-0.8958797346140276 -0.45814536593707755 -0.1438410362258905; 0.1438410362258904 0.45814536593707755 0.8958797346140273]
-@test normalize_log10(v1) == [0.47712125471966244, 0.6020599913279624, 0.6989700043360189, 0.7781512503836436, 0.8450980400142568]
-@test normalize_log10(m1, bych=true) == [0.47712125471966244 0.6020599913279624 0.6989700043360189; 0.9542425094393249 1.0 1.0413926851582251]
-@test normalize_log10(m1, bych=false) == [0.47712125471966244 0.6020599913279624 0.6989700043360189; 0.7781512503836436 0.8450980400142568 0.9030899869919435]
-@test normalize_neglog(v1) == [-0.0, -0.6931471805599453, -1.0986122886681098, -1.3862943611198906, -1.6094379124341003]
-@test normalize_neglog(m1, bych=true) == [-0.0 -0.6931471805599453 -1.0986122886681098; -1.3862943611198906 -1.6094379124341003 -1.791759469228055]
[…]
