### 2025-03-13 abda3f3
fix: list_component()
