### 2025-03-13 eb9d33e
add: format info/warnings using Crayons
