+++ Project.toml
@@ -10,6 +10,7 @@ CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 Cairo = "159f3aea-2a34-519c-b102-8c37f9878175"
 ColorSchemes = "35d6a980-a343-548e-a6ea-1d62b119f2f4"
 ContinuousWavelets = "96eb917e-2868-4417-9cb6-27e7ff17528f"
+Crayons = "a8cc5b0e-0ffa-5ad4-8c14-923d3ee1735f"
 CubicSplines = "9c784101-8907-5a6d-9be6-98f00873c89b"
 DICOM = "a26e6606-dd52-5f6a-a97f-4f611373d757"
 DSP = "717857b8-e6f2-59f4-9121-6e50c889abd2"
@@ -79,6 +80,7 @@ CUDA = "5.5"
 Cairo = "1.1"
 ColorSchemes = "3"
 ContinuousWavelets = "1"
+Crayons = "4.1.1"
 CubicSplines = "0.2"
 DICOM = "0.11"
 DSP = "0.7, 0.8"
