### 2025-03-14 1919494
update: tpt_detect()
