+++ Project.toml
@@ -114,7 +114,7 @@ MAT = "0.10"
 MLJ = "0.20"
 MultivariateStats = "0.10"
 NPZ = "0.4"
-NeuroStats = "0.1.0"
+NeuroStats = "0.2.2"
 PiGPIO = "0.2"
 Plots = "1.40"
 Polynomials = "3.2, 4"
