### 2025-03-14 dc17b8d
update: detect_bad()
