+++ test/plots.jl
@@ -223,7 +223,7 @@ p = plot_mep(mep, ch=["MEP1", "MEP2"], type=:stack)
 @info "Test: plot_ci()"
 s = eeg.data[1, 1:100, :]
 t = eeg.epoch_time[1:100]
-s_avg, s_l, s_u = NeuroStats.bootstrap_ci(s, ci=0.95)
+s_avg, s_l, s_u = NeuroStats.bootstrap_ci(s, cl=0.95)
 p = plot_ci(s_avg, s_l, s_u, t)
 @test p isa Plots.Plot{Plots.GRBackend}
 
