### 2025-03-15 #7734 @AdamWysokinski
@Kevin-Mattheus-Moerman

Double check authors and affiliations (including ORCIDs)
Done.

Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
Current stable version is: 0.25.3

Archive the release on Zenodo/figshare/etc and post the DOI here.
DOI: 10.5281/zenodo.15033084

Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
Checked.

Make sure that the license listed for the archive is the same as the software license.
Checked.

In the paper you list works that use your software. Instead of listing those 5 papers, please cite them so they appear in full in the reference list.
Corrected.

Check if you should add an acknowledgements section to the paper (e.g. to mention funding), if needed you can still add one now.
Checked.

For the Julia paper you cite the old Arxiv pdf, perhaps check if citing this work (what I think is the final paper) would be better: https://doi.org/10.1137/141000671
Corrected.
