### 2025-03-16 #7734 @AdamWysokinski
> [@AdamWysokinski](https://github.com/AdamWysokinski) I think this DOI warning relates to the Arxiv link you cite as well as the final paper. If that is intentional that is fine and no action is needed.

It seems there is no other DOI for the paper. I accept it.
