### 2025-03-16 #7734 @Kevin-Mattheus-Moerman
@AdamWysokinski just to explain on that DOI, it is picking up that you cite both the Arxiv version AND the paper version. SEe screenshot below, the orange highlighted in the bottom is the citation to the Arxiv link while the top citation is for the paper which has the same title. The DOI for the paper cited is fine, but there is no DOI for the Arxiv version. If you meant to only cite the paper then replace the other Arxiv occurrence with the paper citation. If you do want to cite both e.g. the earlier Arxiv version because you say "from the beginning" (and it was published much earlier than final paper), then this is fine. Just checking.

![Image](https://github.com/user-attachments/assets/63bea90d-1909-45ba-9f5a-cac301e39e07)
