### 2025-03-16 #7734 @Kevin-Mattheus-Moerman
@AdamWysokinski

- [ ] Make sure the archive title fully matches the paper title, it currently only says `NeuroAnalyzer`
