### 2025-03-16 #7734 @AdamWysokinski
> [@AdamWysokinski](https://github.com/AdamWysokinski)
>
> * [ ]  Make sure the archive title fully matches the paper title, it currently only says `NeuroAnalyzer`

Shall I rename the archive?
