### 2025-03-16 #7734 @AdamWysokinski
Ok, but: should I change the title of the Zenodo entry to "NeuroAnalyzer: Julia toolbox for analyzing neurophysiological data"?
