### 2025-03-16 #7734 @AdamWysokinski
If so, it's done: https://zenodo.org/records/15033084
