### 2025-03-16 #7734 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
