### 2025-03-16 #7734 @Kevin-Mattheus-Moerman
@AdamWysokinski as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. Below are some final checks:
