### 2025-03-16 c6aa6e9
update: JOSS paper
